### 2022-06-21 #177 @whedon
[PRE REVIEW]: The Data Behind Dark Matter: Exploring Galactic Rotation
**Submitting author:** @villaa (<a href="http://orcid.org/0000-0002-3893-7259">Anthony Villano</a>)
**Repository:** <a href="https://github.com/villano-lab/galactic-spin-W1" target ="_blank">https://github.com/villano-lab/galactic-spin-W1</a>
**Version:** v3.0.0
**Managing EiC:** Jordan Gorzalski

Due to the challenges of the COVID-19 pandemic, JOSE is currently operating in a "reduced service mode". You can read more about what that means in [our blog post](https://blog.joss.theoj.org/2020/05/reopening-joss).
