### 2022-06-21 8e17b86
Remove NGC5533_functions library First stage of library condensing is complete!
