+++ binder/python/components.py
@@ -6,16 +6,173 @@ import numpy as np
 import dataPython as dp
 import inspect
 import types
+import sys
 import scipy.integrate as si
+import scipy.optimize as so
+import scipy.special as ss
+import scipy.interpolate as inter
 from scipy.interpolate import InterpolatedUnivariateSpline      # Spline function
 import lmfit as lm                                              # Fitting
 #custom libraries
 import load_galaxies as galdata
-import NGC5533_functions as funcNGC5533
 
+try:
+    import h5py as h5
+    h5py = 1
+except ModuleNotFoundError:
+    h5py = 0
+    print("Could not find h5py. Datasets will not be able to be saved or loaded using components.py.")
+
+defaultpath = '../'
 ngc7814list = ['ngc7814', 'ngc 7814', '7814']
 ngc5533list = ['ngc5533', 'ngc 5533', '5533']
 
+################################
+########## Constants ###########
+################################
+#Defaults based on NGC5533
+
+#---------Definitely Constant---------
+G = 4.30091e-6                           # Gravitational constant (kpc/solar mass*(km/s)^2)
+#rhocrit = 9.3e-18                        # Critical density of the Universe (kg/km^3). 
+
+#---------Measured Directly-----------
+L = 3.27e10                              # Luminosity (Solar Luminosities)
+absmag = -22.02                          # Absolute magnitude
+magsun = 4.42                            # Absolute magnitude of the sun
+L0 = np.power(10, (0.4*(magsun-absmag))) # Absolute magnitude to luminosity
+
+#---------Measured Indirectly---------
+ups = 2.8                         # Bulge mass-to-light ratio (Solar Mass/Solar Luminosity). Source: Noordermeer, 2008
+q = 0.33                          # Intrinsic axis ratio. Source: Noordermeer, 2008
+e2 = 1-(q**2)                     # Eccentricity. Source: Noordermeer, 2008
+i = 52*(np.pi/180)                # Inclination angle. Source: Noordermeer & Van Der Hulst, 2007
+h_rc = 1.4                        # Core radius (kpc). Source: Noordermeer, 2008
+#c = 1e-12                         #(what does this constant do?)
+#Mvir = 1e11*((c/(11.7))**(-40/3)) # Virial mass (in solar mass) solved from eq(5)
+Mbh_def = 2.7e9                   # Black Hole mass (in solar mass). Source: Noordermeer, 2008
+
+#---------Definitely Variable---------
+n_c = 2.7                         # Concentration parameter. Source: Noordermeer & Van Der Hulst, 2007
+h_c = 8.9                         # Radial scale-length (kpc). Source: Noordermeer & Van Der Hulst, 2007
+hrho00_c = 0.31e9                 # Halo central surface density (solar mass/kpc^2)
+drho00_c = 0.31e9                 # Disk central surface density (solar mass/kpc^2)
+
+#---------Uncategorized-------------------
+re_c = 2.6                        # Effective radius (kpc). Source: Noordermeer & Van Der Hulst, 2007
+upsdisk = 5.0                     # Disk mass-to-light ratio. Source: Noordermeer, 2008
+#rs = (1/c)*(((3*Mvir)/((4*np.pi*100*rhocrit)))**(1/3))   # Scale radius (kpc)
+#rho_s = (100/3)*((c**3)/(np.log(1+c)-(c/(1+c))))*rhocrit # Characteristic density
+h_gamma = 0
+
+################################
+########### Saving #############
+################################
+
+def savedata(xvalues,yvalues,group,dataset,path=defaultpath,file='Inputs.hdf5'): 
[…]
