### 2022-06-22 bd0bb0b
Remove outdated documentation Information was either contained elsewhere or no longer true.
