### 2022-06-22 1e1f8f8
Condense bulge_library into other libraries for component handling. Also fixed a bug with 'Reveal!' button.
