+++ binder/python/load_galaxies.py
@@ -17,8 +17,8 @@ import matplotlib.pyplot  as plt
 ###############################
 
 NGC5533 = {
-
     # Load data from files for Noordermeer 2008 band and fitted curves
+    
     # 'raw' in the sense that I haven't split everything up. Idk there's probably a better way to name these
     'raw_total'        : dp.getXYdata('data/NGC5533/noord-120kpc-total.txt'     ),
     'raw_blackhole'    : dp.getXYdata('data/NGC5533/noord-120kpc-blackhole.txt' ),
@@ -28,10 +28,15 @@ NGC5533 = {
     'raw_gas'          : dp.getXYdata('data/NGC5533/noord-120kpc-gas.txt'       ),
     'raw_band_btm'     : dp.getXYdata('data/NGC5533/noord-120kpc-bottomband.txt'),
     'raw_band_top'     : dp.getXYdata('data/NGC5533/noord-120kpc-topband.txt'   ),
-
+    
     # Get data from 100kpc file
-    'measured_data'    : dp.getXYdata_wXYerr('data/NGC5533/100kpc_data.txt')
-        
+    'measured_data'    : dp.getXYdata_wXYerr('data/NGC5533/100kpc_data.txt'),
+    
+    # Some constants
+    'i'                : 52,    # Inclination angle [degrees] (Fraternali, Sancisi, and Kamphuis, 2011)  
+    'D_Mpc'            : 54.3,  # Distance [Mpc]
+    'Mabs'             : -21.66,# B-band absolute magnitude of NGC 5533 (Norrdermeer, Van Der Hulst, 2007)
+    'check'            : True
 }
 
 # Parameters ########################
@@ -80,10 +85,28 @@ NGC5533['blackhole']['spline'] = inter.BSpline(NGC5533['blackhole']['t'], NGC553
 # Bulge #############################
 NGC5533['bulge'] = {
     'r' : np.asarray(NGC5533['raw_bulge']['xx']),
-    'v' : np.asarray(NGC5533['raw_bulge']['yy'])
+    'v' : np.asarray(NGC5533['raw_bulge']['yy']),
+    'n'      : 2.7,     # Concentration parameter [unitless] (Noordermeer, Van Der Hulst, 2007)
+    'q'      : 0.33,    # Intrinsic axis ratio [unitless] (Noordermeer, 2008)
+    're_arcsec' : 9.9,     # Effective radius [arcsec] (Noordermeer, Van Der Hulst, 2007)
+    'Lb'     : 3.6e10,  # Bulge luminosity [solar luminosity] (Calculated from absolute magnitude of the bulge)
+    'ML'     : 2.8      # Mass-to-light ratio of bulge [unitless] (Noordermeer, 2008)
 }
 NGC5533['bulge']['t'], NGC5533['bulge']['c'], NGC5533['bulge']['k'] = inter.splrep(NGC5533['bulge']['r'], NGC5533['bulge']['v'])
 NGC5533['bulge']['spline'] = inter.BSpline(NGC5533['bulge']['t'], NGC5533['bulge']['c'], NGC5533['bulge']['k'])
+NGC5533['bulge']['re_rad'] = NGC5533['bulge']['re_arcsec'] * (np.pi / (3600*180)) #arcsec to rad
+NGC5533['D_kpc'] = NGC5533['D_Mpc'] * 1000 #Mpc to kpc
+NGC5533['bulge']['re_kpc'] = NGC5533['bulge']['re_rad'] * NGC5533['D_kpc'] #Effective radius kpc
+NGC5533['sources'] = {
+    'Mabs': ' [unitless] (Source #2, Table A4)',
[…]
