### 2022-06-22 8430961
Work on unifying treatment of galaxies within components.py
