+++ binder/python/components.py
@@ -14,7 +14,7 @@ import scipy.interpolate as inter
 from scipy.interpolate import InterpolatedUnivariateSpline      # Spline function
 import lmfit as lm                                              # Fitting
 #custom libraries
-import load_galaxies as galdata
+from load_galaxies import *
 
 try:
     import h5py as h5
@@ -24,21 +24,17 @@ except ModuleNotFoundError:
     print("Could not find h5py. Datasets will not be able to be saved or loaded using components.py.")
 
 defaultpath = '../'
-ngc7814list = ['ngc7814', 'ngc 7814', '7814']
-ngc5533list = ['ngc5533', 'ngc 5533', '5533']
-ngc0891list = ['ngc0891', 'ngc 0891', '0891', 'ngc891', 'ngc 891', '891']
+
+# For parsing any string into a single possible value for galaxy
+import string
+non_numeric_chars = string.printable[10:]
 
 ################################
 ########## Constants ###########
 ################################
 #Pull from load_galaxies.py dict
-def dict(galaxy):
-    if galaxy.lower() in ngc7814list:
-        return galdata.NGC7814
-    if galaxy.lower() in ngc5533list:
-        return galdata.NGC5533
-    if galaxy.lower() in ngc0891list:
-        return galdata.NGC0891
+def galdict(galaxy):
+    return globals()[galaxy.upper()]        
 
 #Defaults based on NGC5533
 
@@ -191,54 +187,9 @@ def checkfile(group='all',path=defaultpath,file='Inputs.hdf5'):
 # (the r array of the raw data)
 
 # Datapoints:
-def data(galaxy):
-    if galaxy.lower() in ngc7814list:
-        return dp.getXYdata_wYerr('data/NGC7814/7814_measured.dat')
-    elif galaxy.lower() in ngc5533list:
-        return galdata.data
-def data_total(galaxy):
-    if galaxy.lower() in ngc5533list:
-        return galdata.data_total
-    else:
-        return None
-def r_dat(galaxy):
-    if galaxy.lower() in ngc7814list:
-        return np.asarray(data(galaxy)['xx'])
-    elif galaxy.lower() in ngc5533list:
-        return galdata.NGC5533['m_radii']
-def v_dat(galaxy):
-    if galaxy.lower() in ngc7814list:
-        return np.asarray(data(galaxy)['yy'])
-    elif galaxy.lower() in ngc5533list:
-        return galdata.NGC5533['m_velocities']
 def v_err0(galaxy):
     if galaxy.lower() in ngc5533list:
         return galdata.v_err0
-    else:
-        return None
-def v_err1(galaxy):
-    if galaxy.lower() in ngc7814list:
-        return np.asarray(data(galaxy)['ey'])
-    elif galaxy.lower() in ngc5533list:
-        return galdata.NGC5533['m_v_errors']
-    else:
-        raise ValueError('Galaxy string', galaxy, 'not recognized.')
-    
-def disk_raw(galaxy):
-    if galaxy.lower() in ngc7814list:
-        return np.asarray(dp.getXYdata('data/NGC7814/7814_gDisk.dat')['yy'])
-    else:
-        return None
-
-def bulge_raw(galaxy):
-    if galaxy.lower() in ngc7814list:
-        return np.asarray(dp.getXYdata('data/NGC7814/7814_gBulge.dat')['yy'])
-    else:
-        return None
-    
-def gas_raw(galaxy):
-    if galaxy.lower() in ngc7814list:
[…]
