+++ binder/python/widget_BH.py
@@ -93,7 +93,7 @@ layout = {'width':'800px'}
 ### NGC 5533 ###
 def f5533(arraysize,massMiniBH,rcut):
     
-    _,fit_dict = comp.bestfit(comp.totalvelocity_miniBH,'5533')
+    _,fit_dict = comp.bestfit(comp.totalvelocity_miniBH,'NGC5533')
     
     bpref = fit_dict['bpref']
     dpref = fit_dict['dpref']
@@ -101,7 +101,9 @@ def f5533(arraysize,massMiniBH,rcut):
     Mbh = fit_dict['Mbh']
     
     # Define radius
-    r = np.linspace(np.min(comp.r_dat('5533')),np.max(comp.r_dat('5533')),500)
+    r = np.linspace(
+        np.min(comp.galdict('NGC5533')['m_radii']), 
+        np.max(comp.galdict('NGC5533')['m_radii']),500)
     
     # Change input to an integer
     arraysize = int(arraysize)     # units: dot
@@ -144,14 +146,15 @@ def f5533(arraysize,massMiniBH,rcut):
     # Second plot - rotation curve   
     ax2.plot(r,comp.halo_BH(r,scale_5533,arraysize,massMiniBH,rcut),
              label=("Dark Matter Halo - Tiny Black Holes"),color='green')
-    ax2.errorbar(comp.r_dat('5533'),comp.v_dat('5533'),yerr=comp.v_err1('5533'),fmt='bo',label='Data')
-    ax2.plot(r,comp.blackhole(r,Mbh,'5533'),label=("Central Black Hole"),color='black')
-    ax2.plot(r,comp.bulge(r,bpref,'5533'),label=("Bulge"),color='orange')
-    ax2.plot(r,comp.disk(r,dpref,'5533'),label=("Disk"),color='purple')
-    ax2.plot(r,comp.gas(r,gpref,'5533'),label=("Gas"),color='blue')
+    ax2.errorbar(comp.galdict('NGC5533')['m_radii'],comp.galdict('NGC5533')['m_velocities'],
+                 yerr=comp.galdict('NGC5533')['m_v_errors'],fmt='bo',label='Data')
+    ax2.plot(r,comp.blackhole(r,Mbh,'NGC5533'),label=("Central Black Hole"),color='black')
+    ax2.plot(r,comp.bulge(r,bpref,'NGC5533'),label=("Bulge"),color='orange')
+    ax2.plot(r,comp.disk(r,dpref,'NGC5533'),label=("Disk"),color='purple')
+    ax2.plot(r,comp.gas(r,gpref,'NGC5533'),label=("Gas"),color='blue')
     ax2.plot(r,comp.totalvelocity_miniBH(r,scale_5533,arraysize,massMiniBH,rcut,
                                   bpref,dpref,gpref,
-                                  Mbh,'5533'),label=("Total Curve"),color='red')
+                                  Mbh,'NGC5533'),label=("Total Curve"),color='red')
     ax2.set_title('NGC 5533',fontsize=40)
     ax2.set_ylabel('Velocity [km/s]',fontsize=25)
     ax2.set_xlabel('Radius [kpc]',fontsize=25)
@@ -162,12 +165,14 @@ def f5533(arraysize,massMiniBH,rcut):
     ax2.legend(bbox_to_anchor=(1,1), loc="upper left", fontsize=20) 
 
     # Residuals
-    residuals = comp.v_dat('5533') - comp.totalvelocity_miniBH(comp.r_dat('5533'),scale_5533,arraysize,massMiniBH,rcut,
+    residuals = comp.galdict('NGC5533')['m_velocities'] 
+    - comp.totalvelocity_miniBH(comp.galdict('NGC5533')['m_radii'],
[…]
