### 2022-06-22 0bf9171
Progress on consistent treatment in components.py
