### 2022-06-23 af525da
Clean up treatment of strings and luminosities in nb10
