### 2022-06-23 77ee53c
Update saving/loading to allow different luminosities.
