+++ binder/python/components.py
@@ -84,7 +84,7 @@ def savedata(xvalues,yvalues,group,dataset,path=defaultpath,file='Inputs.hdf5'):
         if group.lower() in ['bh','black hole','blackhole']:
             group = 'blackhole'
             print("Group name set to 'blackhole'.")
-        if group in ['dm','DM','Dm','Dark Matter','Dark matter','dark matter','h','H','Halo','darkmatter','Darkmatter','DarkMatter']:
+        if group.lower() in ['dm','dark matter','h','halo','darkmatter']:
             group = 'halo'
             print("Group name set to 'halo'.")
         if group.lower() in ['b','bulge']:
@@ -97,25 +97,24 @@ def savedata(xvalues,yvalues,group,dataset,path=defaultpath,file='Inputs.hdf5'):
             grp = saved.create_group(group)
             grp.create_dataset(dataset,data=[xvalues,yvalues])
         except ValueError:
-            try:
-                grp = saved[group]
-                grp.create_dataset(dataset,data=[xvalues,yvalues])
-            except RuntimeError:
-                x = loaddata(group,dataset,path,file)[0]
-                x = np.append(x,xvalues)
-                y = loaddata(group,dataset,path,file)[1]
-                y = np.append(y,yvalues)
-                x, y = (list(a) for a in zip(*sorted(zip(x, y))))
-                i = 0
-                while i < len(x)-1:
-                    if x[i+1] == x[i]:
-                        x = np.delete(x,i+1)
-                        y = np.delete(y,i+1)
-                    else:
-                        i += 1
-                del grp[dataset]
-                savedata(x,y,group,dataset,path,file)
-                return y
+            grp = saved[group]
+            grp.create_dataset(dataset,data=[xvalues,yvalues])
+        except RuntimeError:
+            x = loaddata(group,dataset,path,file)[0]
+            x = np.append(x,xvalues)
+            y = loaddata(group,dataset,path,file)[1]
+            y = np.append(y,yvalues)
+            x, y = (list(a) for a in zip(*sorted(zip(x, y))))
+            i = 0
+            while i < len(x)-1:
+                if x[i+1] == x[i]:
+                    x = np.delete(x,i+1)
+                    y = np.delete(y,i+1)
+                else:
+                    i += 1
+            del grp[dataset]
+            savedata(x,y,group,dataset,path,file)
+            return y
         finally: #No matter what,
             saved.close()
         #print("Saved.") #Convenient for debugging but annoying for fitting.
@@ -214,24 +213,25 @@ def blackhole(r,M,load=False,save=False):
 def b_gammafunc(x,n=n_c):
     return ss.gammainc(2*n,x)*ss.gamma(2*n)-0.5*ss.gamma(2*n)
 b_root = so.brentq(b_gammafunc,0,500000,rtol=0.000001,maxiter=100) #come within 1% of exact root within 100 iterations
-def b_I0(galaxy,n=n_c,re=re_c):
-    return galdict(galaxy)['bulge']['Lb']*(b_root**(2*n))/(re**2*2*np.pi*n*ss.gamma(2*n))
+def b_I0(L,n=n_c,re=re_c):
+    return L*(b_root**(2*n))/(re**2*2*np.pi*n*ss.gamma(2*n))
 def b_r0(n=n_c,re=re_c):
     return re/np.power(b_root,n)
 def b_innerintegral(m,n=n_c,re=re_c):
[…]
