### 2022-06-23 7382d5e
Move some code outside of if statements Still can't figure out how to get a unified treatment but this is an improvement.
