### 2022-06-24 ca962d4
Update release notes and other release docs. Also fixed a few small formatting things in release_checklist.
