### 2022-06-26 0951bb1
[JB] fixed "click all cells below" issue by adding a button, updated environments with ipylab
