### 2022-06-27 8d6d7b9
Start setting up to use repo2docker through circleci Also found a nice binder badge generator for PRs.
