+++ .circleci/config.yml
@@ -1,3 +1,4 @@
+#based on the one provided by repo2docker
 version: 2
 jobs:
 
