### 2022-06-28 6a05b4d
remove use of sudo
already seem to be running as root
