### 2022-06-29 6a9fb24
Replace deprecated circleci image with a newer one
