+++ .circleci/config.yml
@@ -143,7 +143,7 @@ jobs:
 
   deploy:
     machine:
-      image: circleci/classic:201711-01
+      image: cimg/base:2020.01
     working_directory: /tmp/src
     steps:
       - attach_workspace:
