### 2022-06-29 14adde1
Try a newer image from a different docs page
