+++ .circleci/config.yml
@@ -143,7 +143,7 @@ jobs:
 
   deploy:
     machine:
-      image: cimg/base:2020.01
+      image: cimg/base:2021.04
     working_directory: /tmp/src
     steps:
       - attach_workspace:
