### 2022-06-29 c0e957e
Last try replacing docker image for update_cache
This seems to be from this month, if it isn't found then that isn't the problem.
