### 2022-06-29 c84acdc
Try an image from machine list
Found a proper list here:
https://circleci.com/docs/2.0/configuration-reference#available-linux-machine-images
