### 2022-06-29 af8dbfe
Try to add a run step for nb test
