### 2022-06-29 f77fab4
Revert "Try to add a run step for nb test"
This reverts commit af8dbfeaf1eb72ebfd7c4b6153eaa08fa19a9ebf.
Not working!
