### 2022-06-29 10dc680
deploy on all branches at least until we have deployment on master.
