+++ .circleci/config.yml
@@ -198,6 +198,3 @@ workflows:
       - deploy:
           requires:
             - build
-          filters:
-            branches:
-              only: master
