+++ .circleci/config.yml
@@ -180,7 +180,7 @@ jobs:
             fi
   test:
     docker:
-      - image: "${CONTAINER_NAME}"
+      - image: "${CONTAINER_NAME}:${CIRCLE_SHA1}"
     working_directory: ~/binder
     steps:
       - run:
