### 2022-06-29 6b38dc0
appropriately crop CIRCLE_SHA1 \
