+++ .circleci/config.yml
@@ -179,8 +179,11 @@ jobs:
                   docker push ${CONTAINER_NAME}:${DOCKER_TAG}
             fi
   test:
+    machine:
+      environment: 
+        DOCKER_TAG: $(echo $CIRCLE_SHA1 | cut -c1-10)
     docker:
-      - image: "${CONTAINER_NAME}:${CIRCLE_SHA1}"
+      - image: "${CONTAINER_NAME}:${DOCKER_TAG}"
     working_directory: ~/binder
     steps:
       - run:
