### 2022-06-29 bc537ad
try to fix dockertag showing up blank
