+++ .circleci/config.yml
@@ -181,7 +181,7 @@ jobs:
   test:
     machine:
       environment: 
-        DOCKER_TAG: $(echo $CIRCLE_SHA1 | cut -c1-10)
+        DOCKER_TAG: $(echo "${CIRCLE_SHA1}" | cut -c1-10)
     docker:
       - image: "${CONTAINER_NAME}:${DOCKER_TAG}"
     working_directory: ~/binder
