### 2022-06-29 d537015
try tag specifier that doesn't use shell commands
