### 2022-06-30 3237bc6
Change context of var DOCKER_TAG
