+++ .circleci/config.yml
@@ -179,9 +179,8 @@ jobs:
                   docker push ${CONTAINER_NAME}:${DOCKER_TAG}
             fi
   test:
-    machine:
-      environment: 
-        DOCKER_TAG: ${CIRCLE_SHA1:0:10}
+    environment: 
+      DOCKER_TAG: ${CIRCLE_SHA1:0:10}
     docker:
       - image: "${CONTAINER_NAME}:${DOCKER_TAG}"
     working_directory: ~/binder
