### 2022-06-30 1c0ce81
Try using latest tag instead of retrieving as before.
