### 2022-06-30 6b505eb
Try using whole SHA1 as tag instead Since I couldn't figure out how to use the one they originally provided.
