+++ .circleci/config.yml
@@ -110,7 +110,7 @@ jobs:
             # If not set, use commit id as tag
             if [ ! -n "${DOCKER_TAG:-}" ]
                 then
-                    DOCKER_TAG=$(echo "${CIRCLE_SHA1}" | cut -c1-10)
+                    DOCKER_TAG=${CIRCLE_SHA1}
             fi
             # If not set, define CONTAINER_NAME
             if [ ! -n "${CONTAINER_NAME:-}" ]
@@ -166,7 +166,7 @@ jobs:
             echo "Looking for Docker deployment options"
             if [ ! -n "${DOCKER_TAG:-}" ]
                 then
-                    DOCKER_TAG=$(echo "${CIRCLE_SHA1}" | cut -c1-10)
+                    DOCKER_TAG=$(echo "${CIRCLE_SHA1}" )
             fi
             # If not set, define CONTAINER_NAME
             if [[ ! -n "${CONTAINER_NAME:-}" ]]
@@ -176,11 +176,11 @@ jobs:
             echo "Container name set to ${CONTAINER_NAME}:${DOCKER_TAG}"
             if [[ -n "$DOCKER_PASS" ]]; then
                   docker login -u $DOCKER_USER -p $DOCKER_PASS
-                  docker push ${CONTAINER_NAME}:${DOCKER_TAG} ${CONTAINER_NAME}:latest
+                  docker push ${CONTAINER_NAME}:${DOCKER_TAG}
             fi
   test:
     docker:
-      - image: "${CONTAINER_NAME}:latest"
+      - image: "${CONTAINER_NAME}:${CIRCLE_SHA1}"
     working_directory: ~/binder
     steps:
       - run:
