### 2022-06-30 b2b8eee
Add --current-env arg to circleci tests.
