### 2022-06-30 984722c
Now add coverage upload to circleci tests Apparently codecov should be able to correctly handle multiple coverage reports from different CIs So this will not interfere with existing reports.
