### 2022-06-30 fc7bcff
Add badges for codecov, circleci Also rearranged badges a little for order consistency.
