+++ docs/source/02_Modules.rst
@@ -61,9 +61,9 @@ Description of the learning modules
 **Understanding rotation curves** 
 
 .. image:: https://mybinder.org/badge_logo.svg
-   :target: https://mybinder.org/v2/gh/villano-lab/galactic-spin-W1/v2.0.0?labpath=binder/01_DM_Rotation_Curve_Intro.ipynb
+   :target: https://mybinder.org/v2/gh/villano-lab/galactic-spin-W1/v2.0.1?labpath=binder/01_DM_Rotation_Curve_Intro.ipynb
 
-`See module 01 on GitHub <https://github.com/villano-lab/galactic-spin-W1/blob/v2.0.0/binder/01_DM_Rotation_Curve_Intro.ipynb>`_
+`See module 01 on GitHub <https://github.com/villano-lab/galactic-spin-W1/blob/v2.0.1/binder/01_DM_Rotation_Curve_Intro.ipynb>`_
 
 Rotation curves are a measure of how fast an object is moving at a certain distance from the center. Three kinds of rotation curves are discussed in this notebook: rigid-body rotation, planetary rotation, and galactic flat rotation. With the help of animations and rotation curve plots, the three cases are compared to gain a better understanding of galactic rotation curves.
 
@@ -76,9 +76,9 @@ Rotation curves are a measure of how fast an object is moving at a certain dista
 **Introduction to Dark Matter**
 
 .. image:: https://mybinder.org/badge_logo.svg
-   :target: https://mybinder.org/v2/gh/villano-lab/galactic-spin-W1/v2.0.0?labpath=binder%2F02_Widget_NGC5533_DMonly.ipynb
+   :target: https://mybinder.org/v2/gh/villano-lab/galactic-spin-W1/v2.0.1?labpath=binder%2F02_Widget_NGC5533_DMonly.ipynb
 
-`See module 02 on GitHub <https://github.com/villano-lab/galactic-spin-W1/blob/v2.0.0/binder/02_Widget_NGC5533_DMonly.ipynb>`_
+`See module 02 on GitHub <https://github.com/villano-lab/galactic-spin-W1/blob/v2.0.1/binder/02_Widget_NGC5533_DMonly.ipynb>`_
 
 Mass in a galaxy can be visualized by plotting its rotation curve. A galactic rotation curve is a measure of how fast the stars and gas move in the galaxy at a certain distance from the center. According to Newton's law of gravitation, objects orbiting the center of gravity should depend on the mass enclosed in the system. However, the theoretical rotation curve of the measured visible matter does not agree with the measured velocities of matter in some of the spiral galaxies. Dark Matter is introduced to account for the "missing matter".
 
@@ -91,9 +91,9 @@ Mass in a galaxy can be visualized by plotting its rotation curve. A galactic ro
 **Plotting measured velocities**
 
 .. image:: https://mybinder.org/badge_logo.svg
-   :target: https://mybinder.org/v2/gh/villano-lab/galactic-spin-W1/v2.0.0?labpath=binder%2F03_Measured_Data_Plotting.ipynb
+   :target: https://mybinder.org/v2/gh/villano-lab/galactic-spin-W1/v2.0.1?labpath=binder%2F03_Measured_Data_Plotting.ipynb
 
-`See module 03 on GitHub <https://github.com/villano-lab/galactic-spin-W1/blob/v2.0.0/binder/03_Measured_Data_Plotting.ipynb>`_
+`See module 03 on GitHub <https://github.com/villano-lab/galactic-spin-W1/blob/v2.0.1/binder/03_Measured_Data_Plotting.ipynb>`_
 
[…]
