### 2022-06-30 44efa05
Start setting up doxygen Seem to have one example function rendering!
