+++ binder/python/components.py
@@ -36,18 +36,36 @@ defaultpath = '../'
 #========= Constants ===========
 #===============================
 
-##Retrieve a dictionary of parameters for the associated galaxy.
-#
-#**Arguments:** `galaxy` (string)
-#
-#**galaxy:** The galaxy's full name, including catalog. Not case-sensitive. Ignores spaces. 
 def galdict(galaxy):
+    """!
+    @brief Retrieve a dictionary of parameters for the associated galaxy.
+
+    @param galaxy [String] The galaxy's full name, including catalog. Not case-sensitive. Ignores spaces. 
+
+    @return Return a dictionary of parameters for the associated galaxy.
+
+    @see For information on the parameters contained in the returned dictionary, see the documentation for `load_galaxies.py`.
+
+    **Example:**
+    @code
+        #Define a function that prints the cutoff radius of any galaxy and returns it
+        def rcut(galaxy):
+            galaxydata = galdict(galaxy) #Retrieve the whole dictionary
+            cutoff = galaxydata["rcut"]
+            print(cutoff)
+            return cutoff
+        #Print and assign the cutoff for NGC 5533
+        rcut5533 = rcut('NGC5533')
+    @endcode
+    """
     return globals()[galaxy.upper().replace(" ","")]        
 
 #Defaults based on NGC5533
 
 #---------Definitely Constant---------
-
+"""!
+    @brief Gravitational constant (4.30091e-6 kpc/solar mass * (km/s)^2)
+"""
 G = 4.30091e-6                           # Gravitational constant (kpc/solar mass*(km/s)^2) 
 
 #---------Measured Indirectly---------
@@ -77,6 +95,23 @@ h_gamma = 0
 ########### Saving #############
 ################################
 
+##Utility function for saving a dataset to hdf5
+#**Arguments:** `xvalues` (arraylike), `yvalues` (arraylike), `group` (string), `dataset` (string), `path` (string, optional), `file` (string, optional)
+#
+#**xvalues:** [arraylike] An array of x-values to be saved to file. Typically, these values will represent radius.
+#
+#**yvalues:** [arraylike] An array of y-values to be saved to file. Typically, these values will represent velocity.
+#
+#**group:** [string] Name of a group within the hdf5 file. Examples: 'disk', 'blackhole', 'halo', 'bulge', 'total'
+#
+#**dataset:** [string] Name of the dataset to be saved. 
+# This should be unique to the data; 
+# a good way to do this is to specify the source for experimental data 
+# or the parameters for theoretical "data".
+#
+#**path:** [string] Relative or absolute filepath of the hdf5 file. Does NOT include the filename. Default: `../`.
+#
+#**file:** [string] Name of the file to be saved. May include part of the path, but keep in mind `path` variable will also be read. Default: `Inputs.hdf5`.
 def savedata(xvalues,yvalues,group,dataset,path=defaultpath,file='Inputs.hdf5'): 
 #this is a dummy filename to enforce ordering; try not to save here except for testing!
     if h5py == 1:
@@ -124,7 +159,20 @@ def savedata(xvalues,yvalues,group,dataset,path=defaultpath,file='Inputs.hdf5'):
     elif h5py == 0:
         print("ERROR: h5py was not loaded.")
         return 1
-    
+
+##Utility function for loading a dataset from hdf5
+#**Arguments:** `group` (string), `dataset` (string), `path` (string, optional), `file` (string, optional)
+#
[…]
