### 2022-07-03 25d5b65
Try using autodoc instead Seems to work better!
