### 2022-07-07 d1f5f71
Remove doxyfile using docstrings(?) instead.
