+++ docs/source/conf.py
@@ -24,7 +24,7 @@ except FileExistsError:
     pass
 os.chdir('source')
 
-def configureDoxyfile(input_dir, output_dir):
+"""def configureDoxyfile(input_dir, output_dir):
     with open('Doxyfile', 'r') as file :
         filedata = file.read()
 
@@ -32,7 +32,7 @@ def configureDoxyfile(input_dir, output_dir):
     filedata = filedata.replace('@DOXYGEN_OUTPUT_DIR@', output_dir)
 
     with open('Doxyfile', 'w') as file:
-        file.write(filedata)
+        file.write(filedata)"""
 
 # Check if we're running on Read the Docs' servers
 read_the_docs_build = os.environ.get('READTHEDOCS', None) == 'True'
@@ -44,8 +44,8 @@ breathe_projects = {
 if read_the_docs_build:
     input_dir = '../../binder/python'
     output_dir = 'dox_build'
-    configureDoxyfile(input_dir, output_dir)
-    subprocess.call('doxygen', shell=True)
+    """configureDoxyfile(input_dir, output_dir)
+    subprocess.call('doxygen', shell=True)"""
     breathe_projects['galactic-spin-W1'] = output_dir + '/xml'
 
 
