### 2022-07-07 fae3c7d
Try just removing the readthedocs flags works locally
