### 2022-07-07 4d97cbf
Try adding something to path in readthedocs.
