### 2022-07-08 d502c7a
Add numpy, scipy to requirements
