+++ docs/requirements.txt
@@ -1,4 +1,5 @@
 ###### Requirements without Version Specifiers ######
 myst-parser
 sphinx_rtd_theme
-breathe
+numpy
+scipy
