### 2022-07-08 8611a92
Re-add lmfit as mock import
