### 2022-07-08 28e4807
Add symlink command to readthedocs.
