+++ docs/source/conf.py
@@ -47,6 +47,7 @@ if read_the_docs_build:
     input_dir = '../../binder/python'
     output_dir = 'dox_build'
     sys.path.append('../../binder/python')
+    os.symlink('../../binder/data','data')
     #configureDoxyfile(input_dir, output_dir)
     #subprocess.call('doxygen', shell=True)
     #breathe_projects['galactic-spin-W1'] = output_dir + '/xml'"""
