### 2022-07-08 d53b05b
Reorganize rst files to make room for new documentation
