### 2022-07-08 441bce0
[JB] condensed bulge function
