### 2022-07-08 ee1f873
[JB] bulge function condensed
