+++ docs/source/06_Release_History.rst
@@ -1,0 +1,21 @@
+===============
+Release history
+===============
+
+* 30.06.2022: Release of `v2.0.1 <https://github.com/villano-lab/galactic-spin-W1/releases/tag/v2.0.1>`_
+     Improved notebook 06 by restoring the ability to "Run all cells below" with a button press instead of JupyterLab UI.
+     Added another CI to test an environment generated more similarly to how Binder generates its environments.
+* 24.06.2022: Release of `v2.0.0 <https://github.com/villano-lab/galactic-spin-W1/releases/tag/v2.0.0>`_
+     Reduced the number of libraries by moving many functions into a shared space.
+     Reduced the number of library functions by taking many similar functions and increasing the number of allowed arguments.
+* 17.06.2022: Release of `v1.0.4 <https://github.com/villano-lab/galactic-spin-W1/releases/tag/v1.0.4>`_
+     Author name update.
+* 17.06.2022: Release of `v1.0.3 <https://github.com/villano-lab/galactic-spin-W1/releases/tag/v1.0.3>`_
+     Address author comments regarding paper.
+* 09.06.2022: Release of `v1.0.2 <https://github.com/villano-lab/galactic-spin-W1/releases/tag/v1.0.2>`_
+     Expand and elaborate on documentation.
+     Fix various documentation issues.
+* 27.04.2022: Release of `v1.0.1 <https://github.com/villano-lab/galactic-spin-W1/releases/tag/v1.0.1>`_
+     Fixed various documentation issues.
+* 27.04.2022: Release of `v1.0.0 <https://github.com/villano-lab/galactic-spin-W1/releases/tag/v1.0.0>`_
+     Initial release for RaCAS 2022.
