+++ docs/source/08_Contact.rst
@@ -1,0 +1,8 @@
+=================
+Contact & Support
+=================
+
+The authors of *Galactic Spin Workshop 1* are `A.N. Villano <https://github.com/villaa>`_, `J. Bergfalk <https://github.com/juditbergfalk>`_, `K. Harris <https://github.com/nuclearGoblin>`_, `R. Hatami <https://github.com/mabs33>`_, `F. Vititoe <https://github.com/frankie-768>`_, and `J. Johnston <https://github.com/joohlz>`_.
+
+For questions, support, bug reports, or other suggestions, please open an issue
+on `github <https://github.com/villano-lab/galactic-spin-W1/issues>`_.
