### 2022-07-11 26a7e13
[JB] added descriptions to functions...in progress
