+++ binder/python/components.py
@@ -20,7 +20,8 @@ import scipy.special as ss
 import scipy.interpolate as inter
 from scipy.interpolate import InterpolatedUnivariateSpline      # Spline function
 import lmfit as lm                                              # Fitting
-#custom libraries
+
+# Custom libraries
 from load_galaxies import *
 
 try:
@@ -51,23 +52,24 @@ def galdict(galaxy):
         For information on the parameters contained in the returned dictionary, see the documentation for `load_galaxies.py`.
 
     Example:
-        >>> #Define a function that prints the cutoff radius of any galaxy and returns it
+        >>> # Define a function that prints the cutoff radius of any galaxy and returns it
         >>> def rcut(galaxy):
         >>>     galaxydata = galdict(galaxy) #Retrieve the whole dictionary
         >>>     cutoff = galaxydata["rcut"]
         >>>     print(cutoff)
         >>>     return cutoff
-        >>> #Print and assign the cutoff for NGC 5533
+        >>> # Print and assign the cutoff for NGC 5533
         >>> rcut5533 = rcut('NGC5533')
 
     """
+    
     return globals()[galaxy.upper().replace(" ","")]        
 
-#Defaults based on NGC5533
+# Defaults based on NGC5533
 
 #---------Definitely Constant---------
 ## @brief Gravitational constant (4.30091e-6 kpc/solar mass * (km/s)^2)
-G = 4.30091e-6                           # Gravitational constant (kpc/solar mass*(km/s)^2) 
+G = 4.30091e-6                    # Gravitational constant (kpc/solar mass*(km/s)^2) 
 
 #---------Measured Indirectly---------
 ups = 2.8                         # Bulge mass-to-light ratio (Solar Mass/Solar Luminosity). Source: Noordermeer, 2008
@@ -75,21 +77,17 @@ q = 0.33                          # Intrinsic axis ratio. Source: Noordermeer, 2
 e2 = 1-(q**2)                     # Eccentricity. Source: Noordermeer, 2008
 i = 52*(np.pi/180)                # Inclination angle. Source: Noordermeer & Van Der Hulst, 2007
 h_rc = 1.4                        # Core radius (kpc). Source: Noordermeer, 2008
-#c = 1e-12                         #(what does this constant do?)
-#Mvir = 1e11*((c/(11.7))**(-40/3)) # Virial mass (in solar mass) solved from eq(5)
 Mbh_def = 2.7e9                   # Black Hole mass (in solar mass). Source: Noordermeer, 2008
 
 #---------Definitely Variable---------
 n_c = 2.7                         # Concentration parameter. Source: Noordermeer & Van Der Hulst, 2007
 h_c = 8.9                         # Radial scale-length (kpc). Source: Noordermeer & Van Der Hulst, 2007
-hrho00_c = 0.31e9                 # Halo central surface density (solar mass/kpc^2)
+hrho00_c = 0.31e9                 # Halo central surface density (solar mass/kpc^2). Source: Noordermeer, 2008
 drho00_c = 0.31e9                 # Disk central surface density (solar mass/kpc^2)
 
-#---------Uncategorized-------------------
+#---------Uncategorized---------------
 re_c = 2.6                        # Effective radius (kpc). Source: Noordermeer & Van Der Hulst, 2007
 upsdisk = 5.0                     # Disk mass-to-light ratio. Source: Noordermeer, 2008
-#rs = (1/c)*(((3*Mvir)/((4*np.pi*100*rhocrit)))**(1/3))   # Scale radius (kpc)
-#rho_s = (100/3)*((c**3)/(np.log(1+c)-(c/(1+c))))*rhocrit # Characteristic density
 h_gamma = 0
 
 ################################
[…]
