### 2022-07-12 93891a4
[JB] added documentation for functions
