+++ binder/python/components.py
@@ -20,6 +20,7 @@ import scipy.special as ss
 import scipy.interpolate as inter
 from scipy.interpolate import InterpolatedUnivariateSpline      # Spline function
 import lmfit as lm                                              # Fitting
+import traceback
 
 # Custom libraries
 from load_galaxies import *
@@ -60,7 +61,6 @@ def galdict(galaxy):
         >>>     return cutoff
         >>> # Print and assign the cutoff for NGC 5533
         >>> rcut5533 = rcut('NGC5533')
-
     """
     
     return globals()[galaxy.upper().replace(" ","")]        
@@ -257,11 +257,14 @@ def loaddata(group,
         dset = grp[dataset]
         a = dset[:]
         return a
+    
     # Placeholder; I will design this to store information at a later date.
     elif h5py == 0:
         print("ERROR: h5py was not loaded.")
         return 1
-    saved.close() # No matter what, close the file when you're done
+    
+    # No matter what, close the file when you're done
+    saved.close() 
     
 ##Utility function for checking data present in hdf5 without loading.
 #**Arguments:** `group` (string, optional), `path` (string, optional), `file` (string, optional)
@@ -321,6 +324,7 @@ def checkfile(group='all',
             for n in grp:
                 print(grp[n])
         saved.close()
+        
     elif h5py == 0:
         print("ERROR: h5py was not loaded.")
         return 1
@@ -365,9 +369,9 @@ def blackhole(r,
 
     Parameters:
         r : [arraylike] 
-            Radius values or distance from the center of the galaxy used to calculate velocities, in kpc.
+            Radius values or distance from the center of the galaxy used to calculate velocities (in kpc).
         M : [float] 
-            Mass of the black hole, in solar masses.
+            Mass of the black hole (in solar masses).
         load : [bool] 
             Whether or not to load data from a file. If no data can be loaded, it will be saved for future use instead. 
             Default: `False`.
@@ -376,12 +380,12 @@ def blackhole(r,
             Default: `False`.
         
     Returns:
-        An array of rotational velocities.        
+        An array of rotational velocities (in km/s).        
 
     Example:
         >>> # Calculate the gravitational effect of a black hole the size of 1000 suns, 10 kpc away. 
-        >>> print('Velocity of a star 10 kpc away from a 1000 solar mass black hole = {} km/s.'.format(blackhole(10,1000)))
-        >>> Velocity of a star 10 kpc away from a 1000 solar mass black hole = [0.02073864] km/s.  
+        >>> print(blackhole(r=10, M=1000))
+        >>> [0.02073864] 
     """
     
     # Define component for saving
@@ -430,24 +434,23 @@ def bulge(r,
 
     Parameters:
         r : [array]
-            Radius values or distance from the center of the galaxy used to calculate velocities, in kpc.
+            Radius values or distance from the center of the galaxy used to calculate velocities (in kpc).
         bpref : [float]
-            Bulge prefactor or scaling factor, unitless.
+            Bulge prefactor or scaling factor (unitless).
         n : [float]
-            Concentration parameter, unitless. Default: `2.7`.
+            Concentration parameter (unitless). Default: `2.7`.
         re : [float]
-            Effective radius, in kpc.
+            Effective radius (in kpc). Default: `2.6`.
         galaxy : [string]
             The galaxy's full name, including catalog. Not case-sensitive. Ignores spaces. 
 
     Returns:
-        An array of splined bulge velocities.        
+        An array of splined bulge velocities (in km/s).        
 
     Example:
[…]
