### 2022-07-13 13cec4e
[JB] added scale and arraysize to example
