### 2022-07-13 df0b542
[JB] added more documentation to functions
