+++ binder/python/widget_SPARC.py
@@ -70,6 +70,32 @@ distance = float(firstline[3])
 ##############################
 
 def interpd(x,y):
+    """
+    Interpolation function with degree of the smoothing spline of 3, using scipy.interpolate.InterpolatedUnivariateSpline.
+
+    Parameters:
+        x : [array]
+            An array of x-values. 
+        y : [array]
+            An array of y-values. 
+
+    Returns:
+        Polynomial that can be called as a function.
+
+    .. note::
+        The number of data points must be larger than the spline degree k.
+
+    Example:
+        >>> # Define x and y values and interpolate 
+        >>> x = np.linspace(-5, 5, 50)
+        >>> y = 0.2*np.exp(-x**2)
+        >>> spline = interpd(x, y)
+        >>> plt.plot(x, y, 'ro')
+        >>> x_resampled = np.linspace(-5, 5, 1000)
+        >>> plt.plot(x_resampled, spline(x_resampled))
+        >>> plt.show()
+    """
+    
     return InterpolatedUnivariateSpline(x,y,k=3)
 
 ################################
@@ -78,16 +104,83 @@ def interpd(x,y):
 
 # Bulge
 def bulge(r,bpref):
+    """
+    Interpolating the bulge velocity.
+
+    Parameters:
+        r : [array]
+            Sampling radius values or distance from the center of the galaxy (in kpc).           
+        bpref : [float]
+            Bulge prefactor or scaling factor (unitless). 
+
+    Returns:
+        Splined bulge velocity as a function of sampling radii.
+
+    Example:
+        >>> # Define measured radius and velocity and interpolate them
+        >>> Rad = np.array([0, 0.5, 1.2, 2.6, 5.3, 6.7, 7.1, 9.5])
+        >>> Vbul = np.array([100, 200, 280, 200, 130, 120, 110, 100])
+        >>> r = np.linspace(0, 20, 50)
+        >>> bulgespline = bulge(r, bpref=1)
+        >>> plt.plot(Rad, Vbul, 'ro')
+        >>> plt.plot(r, bulgespline)
+        >>> plt.show()
+    """
+    
     polynomial = interpd(Rad,bpref*Vbul)   
     return polynomial(r)
 
 # Disk
 def disk(r,dpref):
+    """
+    Interpolating the disk velocity.
+
+    Parameters:
+        r : [array]
+            Sampling radius values or distance from the center of the galaxy (in kpc).           
+        dpref : [float]
+            Disk prefactor or scaling factor (unitless). 
+
+    Returns:
+        Splined disk velocity as a function of sampling radii.
+
+    Example:
+        >>> # Define measured radius and velocity and interpolate them
+        >>> Rad = np.array([0, 0.5, 1.2, 2.6, 5.3, 6.7, 7.1, 9.5])
+        >>> Vdisk = np.array([100, 200, 280, 200, 130, 120, 110, 100])
+        >>> r = np.linspace(0, 20, 50)
+        >>> diskspline = disk(r, dpref=1)
+        >>> plt.plot(Rad, Vdisk, 'ro')
+        >>> plt.plot(r, diskspline)
+        >>> plt.show()
+    """
+    
     polynomial = interpd(Rad,dpref*Vdisk)   
     return polynomial(r)
 
 # Gas 
 def gas(r):
+    """
+    Interpolating the gas velocity.
+
+    Parameters:
+        r : [array]
+            Sampling radius values or distance from the center of the galaxy (in kpc).
+
+    Returns:
+        Splined gas velocity as a function of sampling radii.
+
+    Example:
+        >>> # Define measured radius and velocity and interpolate them
[…]
