### 2022-07-15 e836b22
Fix badges
Also set up for next release while updating releasenotes
Going to be a major version (3.0.0) as we are removing library API in favor of having fewer micro-functions
