+++ docs/source/06_Release_History.rst
@@ -2,6 +2,10 @@
 Release history
 ===============
 
+* In progress: Release of `v3.0.0 <https://github.com/villano-lab/galactic-spin-W1/releases/tag/v3.0.0>`_
+     Documented library functions.
+     Condensed small library functions that are only called by other library functions into fewer, larger ones.
+     Fix readthedocs badges.
 * 30.06.2022: Release of `v2.0.1 <https://github.com/villano-lab/galactic-spin-W1/releases/tag/v2.0.1>`_
      Improved notebook 06 by restoring the ability to "Run all cells below" with a button press instead of JupyterLab UI.
      Added another CI to test an environment generated more similarly to how Binder generates its environments.
