### 2022-07-15 0248145
Move to using dataPython via pip
Since we use dataPython in many repositories, makes more sense to have it in one centralized place.
Should still document this, but now we can document it everywhere instead of only here.
