### 2022-07-23 050601c
set up to auto-include all functions in docs
next step: get them in ToC.
Saw autoapi extension recommended for this.
