+++ docs/source/03_Glossary.rst
@@ -1,6 +1,6 @@
-========
-Glossary
-========
+===========
+3. Glossary
+===========
 
 .. glossary::
 
