+++ docs/source/04_Library_Functions.rst
@@ -1,3 +1,33 @@
 ====================
 4. Library Functions
 ====================
+
+components
+==========
+.. automodule:: components
+    :members:
+
+downloadSPARCdata
+=================
+.. automodule:: downloadSPARCdata
+    :members:
+
+load_galaxies
+=============
+.. automodule:: load_galaxies
+    :members:
+
+rc_animation
+============
+.. automodule:: rc_animation
+    :members:
+
+widget_BH
+=========
+.. automodule:: widget_BH
+    :members:
+
+widget_SPARC
+============
+.. automodule:: widget_SPARC
+    :members:
