### 2022-07-23 9403e8b
[JB] updated function documentation
