### 2022-07-23 2905497
Add ipywidgets to mock imports.
