### 2022-07-23 c3475a4
Have readthedocs create itself a symlink for build.
