+++ docs/source/index.rst
@@ -43,7 +43,7 @@ If you want to contribute to `galactic-spin-W1`, please check out the `contribut
    01_Getting_Started
    02_Modules
    03_Glossary
-   04_Library_Functions
+   autoapi/index
    05_Citations
    06_Release_History
    07_License
