+++ docs/requirements.txt
@@ -5,3 +5,4 @@ numpy
 scipy
 lmfit
 dataPython
+autoapi
