### 2022-07-24 c10bf9a
Reorganize module documentation with template Also worked on the docs some
