+++ binder/python/widget_SPARC.py
@@ -1,3 +1,7 @@
+"""
+A module for handling the SPARC widget.
+"""
+
 ################################
 ########### Imports ############
 ################################
@@ -15,6 +19,10 @@ from IPython.display import display
 import warnings
 warnings.filterwarnings("ignore")         # ignore warnings
 
+import sys
+sys.path.append('python')
+import components as comp
+
 ################################
 ############ Data ##############
 ################################
@@ -23,46 +31,66 @@ warnings.filterwarnings("ignore")         # ignore warnings
 
 # Read the file where the chosen galaxy name is located (this was modified by the user)
 textfile = open('python/chosengalaxy.txt', 'r')
+"""file: A local text file, opened in read-only mode, indicating a chosen galaxy.
+
+The purpose of this file is to:  
+1. Allow this chosen galaxy variable to be passed flexibly between programs.
+2. Allow this chosen galaxy to be retained between sessions.
+
+.. seealso:: The data in this text file is stored in the :func:`galaxy <widget_SPARC.galaxy>` variable.
+"""
 galaxy = textfile.read()
+"""string: The name of the chosen galaxy.
+
+.. seealso:: This string is retrieved from :func:`textfile <widget_SPARC.textfile>`.
+"""
 textfile.close()
 
 # Define SPARC directory
-SPARC_file_directory='./data/sparc/'                       #note that '' means the string variable is blank
+#SPARC_file_directory='./data/sparc/'                       #note that '' means the string variable is blank
+"""string: The directory containing SPARC files.
+"""
 
 # Define file path for .dat files
-SPARC_file_path = SPARC_file_directory + galaxy + '_rotmod.dat'
+SPARC_file_path = './data/sparc/' + galaxy + '_rotmod.dat'
+"""string: The filename (including relative path) containing data for the :func:`chosen galaxy <galaxy>`.
+"""
 
 # Load the galaxy data
 data = np.loadtxt(SPARC_file_path)
+"""ndarray: A numpy array of data loaded from the :func:`SPARC file path <SPARC_file_path>`.
+"""
 
 # Split columns into arrays
 Rad,Vobs,errV,Vgas,Vdisk,Vbul,SBdisk,SBbul = data.T
 
 # Check if it has all components
 if np.sum(Vbul) == 0:     # Bulge
-    nobulge = True
     warning_bulge = "There is no bulge component."
+    """string: If the bulge component is missing, a message indicating that it is. Otherwise, an empty string.
+    """
 else: 
-    nobulge = False
     warning_bulge = ""
 
 if np.sum(Vdisk) == 0:    # Disk
-    nodisk = True
     warning_disk = "There is no disk component."
+    """string: If the disk component is missing, a message indicating that it is. Otherwise, an empty string.
+    """
 else: 
-    nodisk = False
     warning_disk = ""
 
 if np.sum(Vgas) == 0:     # Gas
-    nogas = True
     warning_gas = "There is no gas component."
+    """string: If the gas component is missing, a message indicating that it is. Otherwise, an empty string.
+    """
 else: 
-    nogas = False
     warning_gas = ""
 
 # Define distance to galaxy in Mpc
-firstline = open(SPARC_file_path).readline().split()
-distance = float(firstline[3])
+with open(SPARC_file_path) as file:
+    distance = float(file.readline().split()[3])
+    """float: Distance to galaxy, in Mpc.
+    """
 
 ##############################
 ### Interpolation function ###
[…]
