### 2022-07-24 89244be
Continue work on documenting. Removing some of the more granular API where it doesn't hurt optimiztion to make docs more navigable.
