### 2022-07-26 49eb945
Remove symlinks as they are not needed by autoapi And they don't play nicely with git.
