+++ docs/source/conf.py
@@ -37,11 +37,8 @@ if read_the_docs_build:
     input_dir = '../../binder/python'
     output_dir = 'dox_build'
     sys.path.append('../../binder/python')
-    os.symlink('../../binder/data','data')
-    os.symlink('../../binder/python','python')
-    #configureDoxyfile(input_dir, output_dir)
-    #subprocess.call('doxygen', shell=True)
-    #breathe_projects['galactic-spin-W1'] = output_dir + '/xml'"""
+    #os.symlink('../../binder/data','data')
+    #os.symlink('../../binder/python','python')
 
 
 # -- Project information -----------------------------------------------------
