### 2022-07-26 52f4b33
Progress on documentation. Reorganized widget_BH a little bit since there were duplicate variables for different galaxies.
