+++ binder/python/widget_BH.py
@@ -19,14 +19,21 @@ import dataPython as dp
 
 # Import galaxy image
 img = plt.imread("images/A_spiral_snowflake.jpg")            # Import image of NGC 6814
+"""[ndarray]: An array of RGB data for a galaxy image (`images/A_spiral_snowflake.jpg`), imported using `matplotlib.pyplot.imread`.
+"""
 
 # Find the center by eye from the image
-c_x = 1960
-c_y = 1800
+center = [1960,1800]
+"""[list]: Coordinate pair indicating the pixel location of the center of the galaxy in :func:`img <widget_BH.img>`.
+"""
 
 # Kpc limits, visual guess based on the galaxy in the image chosen:
 minkpc = 0
+"""[int]: The minimum value, in kpc, allowed for the radial position of black holes on the plot.
+"""
 maxkpc = 100
+"""[int]: The maximum value, in kpc, allowed for the radial position of black holes on the plot.
+"""
 
 ##################
 ### Parameters ###
@@ -34,57 +41,123 @@ maxkpc = 100
 
 # units: scale = [#number of actual black holes / plotted dot]
 kpctopixels = 20                # visual scaling, varies depending on size of galaxy image (and actual size of galaxy)
-r1 = minkpc*kpctopixels
-r2 = maxkpc*kpctopixels
+"""[int]: Number of pixels per kpc for plotting black holes over black hole image.
+"""
 
 # For mass of black hole slider:
 minmassBH = 0.1                 # solar masses, arbitrary
-maxmassBH = 3.8                 # solar masses, size of the smallest black hole ever discovered according to
-                                # https://www.scientificamerican.com/gallery/the-smallest-known-black-hole/
+"""[float]: The minimum value, in solar masses, allowed for the (average) black hole mass.
+"""
+maxmassBH = 3.8                 # https://www.scientificamerican.com/gallery/the-smallest-known-black-hole/
+"""[float]: The maximum value, in solar masses, allowed for the (average) black hole mass.
+"""
 defaultmass = 1.5               # default mass value for slider
-stepM = minmassBH               # step of mass of bh's
+"""[float]: The default value, in solar masses, for the (average) black hole mass.
+"""
+scale = 1e6                     # scale is neccessary to be a constant, otherwise the widget will freeze up the computer! 
+"""[int]: The number of black holes represented by each dot in the galaxy plot.
+"""
+# For number of black holes slider:
+stepN = 5
+"""[int]: The step size of the slider controlling the number of black holes. In terms of number of black holes represented, this step size is multiplied by the :func:`scale <widget_BH.scale>`.
+"""
+def minnumberBH(galaxy):
+    """A function that returns the minimum number of black holes (prior to multiplying by the :func:`scale <widget_BH.scale>`) appropriate for the supplied galaxy.
 
-### NGC 5533 ### 
-# For number of black holes slider
-scale_5533 = 1e6                      # scale is neccessary to be a constant, otherwise the widget will freeze up the computer! 
-                                      # scale is how many black holes represent each dot on image
-minnumberBH_5533 = 50                 # min number of black holes (this is multiplied by the scale)
-maxnumberBH_5533 = 5e2                # max number of black holes (this is multiplied by the scale)
[…]
