### 2022-07-26 dba5ada
First go-through of annotating widget_SPARC
and adding images to examples in rc_animation
