+++ binder/python/rc_animation.py
@@ -8,33 +8,42 @@ def CalculatePosition(radius,velocity,time,dt):
     """
     Calculates positions of multiple objects around a circle after some time with interval dt.
 
+    Any set of units may be used, provided they match -- 
+    for example, if using a radius of meters and time in seconds, velocity should be in meters per second and the time increments should be in seconds.
+
     Parameters:
         radius : [array]
             Radius of the objects from the center. Units to match velocity.
         velocity : [array]
-            Velocity of the objects. Units to match radius and time.
+            Velocity of the objects. Units to match radius and time. Length must match length of radius.
         time : [int]
-            Maximum time. Units to match velocity.
+            Maximum time. Units to match velocity. If not divisible by dt, the final time provided may exceed this value by a number less than dt.
         dt : [int]
             Time increment. Units to match time.
 
     Returns:
-        Two [ndarray]s of x-position, y-position of all objects
-        [1darray] of associated time.
+        Two [ndarray]s of x-position (first array) and y-position (second array) of all objects. These arrays are stored such that each row represents a single time 
+        and each column represents a single object.  
+        [1darray] of associated time.  
 
     Example:
+        >>> import numpy as np #To generate the arrays
         >>> radius = np.array([1,2,3,4,5])              # in m
         >>> velocity = np.array([0.1,0.2,0.3,0.4,0.5])  # in m/s
         >>> time = 100                                  # in s
-        >>> dt = 1                                      # in s
+        >>> dt = 33                                     # in s
         >>> CalculatePosition(radius,velocity,time,dt)
-        >>> (array([[ 1.        ,  2.        ,  3.        ,  4.        ,  5.        ],
-        >>>         [ 0.28366219,  0.56732437,  0.85098656,  1.13464874,  1.41831093],
-        >>>         [-0.83907153, -1.67814306, -2.51721459, -3.35628612, -4.19535765]]),
-        >>>  array([[ 0.        ,  0.        ,  0.        ,  0.        ,  0.        ],
-        >>>         [-0.95892427, -1.91784855, -2.87677282, -3.8356971 , -4.79462137],
-        >>>         [-0.54402111, -1.08804222, -1.63206333, -2.17608444, -2.72010555]]),
-        >>>  array([  0,  50, 100]))
+        (array([[ 1.        ,  2.        ,  3.        ,  4.        ,  5.        ],
+            [-0.98747977, -1.97495954, -2.96243931, -3.94991908, -4.93739885],
+            [ 0.95023259,  1.90046518,  2.85069778,  3.80093037,  4.75116296],
+            [-0.88919115, -1.77838231, -2.66757346, -3.55676461, -4.44595576],
+            [ 0.80588396,  1.61176792,  2.41765187,  3.22353583,  4.02941979]]), 
+        array([[ 0.        ,  0.        ,  0.        ,  0.        ,  0.        ],
+            [-0.15774569, -0.31549139, -0.47323708, -0.63098278, -0.78872847],
+            [ 0.31154136,  0.62308273,  0.93462409,  1.24616545,  1.55770682],
+            [-0.45753589, -0.91507179, -1.37260768, -1.83014358, -2.28767947],
+            [ 0.59207351,  1.18414703,  1.77622054,  2.36829406,  2.96036757]]), 
+        array([  0,  33,  66,  99, 132]))
     """
     
     # Initial conditions
@@ -71,37 +80,41 @@ def MultiplePositions(radius,velocity,time,dt):
     """
     Calculates transposed positions of multiple objects around a circle after some time with interval dt. Setup for the animation.
 
+    Any set of units may be used, provided they match -- 
+    for example, if using a radius of meters and time in seconds, velocity should be in meters per second and the time increments should be in seconds.
+
     Parameters:
         radius : [array]
[…]
