+++ binder/python/widget_SPARC.py
@@ -1,5 +1,5 @@
 """
-A module for handling the SPARC widget.
+A module for handling the SPARC widget, notebook `09_Widget_SPARC_Galaxies.ipynb <https://github.com/villano-lab/galactic-spin-W1/blob/master/binder/09_Widget_SPARC_Galaxies.ipynb>`_.
 """
 
 ################################
@@ -92,39 +92,6 @@ with open(SPARC_file_path) as file:
     """float: Distance to galaxy, in Mpc.
     """
 
-##############################
-### Interpolation function ###
-##############################
-
-def interpd(x,y):
-    """
-    Interpolation function with degree of the smoothing spline of 3, using scipy.interpolate.InterpolatedUnivariateSpline.
-
-    Parameters:
-        x : [array]
-            An array of x-values. 
-        y : [array]
-            An array of y-values. 
-
-    Returns:
-        Polynomial that can be called as a function.
-
-    .. note::
-        The number of data points must be larger than the spline degree k.
-
-    Example:
-        >>> # Define x and y values and interpolate 
-        >>> x = np.linspace(-5, 5, 50)
-        >>> y = 0.2*np.exp(-x**2)
-        >>> spline = interpd(x, y)
-        >>> plt.plot(x, y, 'ro')
-        >>> x_resampled = np.linspace(-5, 5, 1000)
-        >>> plt.plot(x_resampled, spline(x_resampled))
-        >>> plt.show()
-    """
-    
-    return InterpolatedUnivariateSpline(x,y,k=3)
-
 ################################
 ######### Components ###########
 ################################
@@ -141,10 +108,11 @@ def bulge(r,bpref):
             Bulge prefactor or scaling factor (unitless). 
 
     Returns:
-        Splined bulge velocity as a function of sampling radii.
+        Splined bulge velocity as a [function] of sampling radii.
 
     Example:
         >>> # Define measured radius and velocity and interpolate them
+        >>> import numpy as np
         >>> Rad = np.array([0, 0.5, 1.2, 2.6, 5.3, 6.7, 7.1, 9.5])
         >>> Vbul = np.array([100, 200, 280, 200, 130, 120, 110, 100])
         >>> r = np.linspace(0, 20, 50)
@@ -152,9 +120,12 @@ def bulge(r,bpref):
         >>> plt.plot(Rad, Vbul, 'ro')
         >>> plt.plot(r, bulgespline)
         >>> plt.show()
+
+        .. image:: ../../images/bulgeExample.png
+            :alt: A plot of the supplied bulge velocities, with radii as the x-points and velocities as the y-points, and a blue line representing a rough interpolation of them with scaling factor 1.
     """
     
-    polynomial = interpd(Rad,bpref*Vbul)   
+    polynomial = InterpolatedUnivariateSpline(Rad,bpref*Vbul,k=3)
     return polynomial(r)
 
 # Disk
@@ -169,20 +140,23 @@ def disk(r,dpref):
             Disk prefactor or scaling factor (unitless). 
 
     Returns:
-        Splined disk velocity as a function of sampling radii.
+        Splined disk velocity as a [function] of sampling radii.
 
     Example:
         >>> # Define measured radius and velocity and interpolate them
-        >>> Rad = np.array([0, 0.5, 1.2, 2.6, 5.3, 6.7, 7.1, 9.5])
-        >>> Vdisk = np.array([100, 200, 280, 200, 130, 120, 110, 100])
+        >>> #Rad = #Plotting Rad and Vdisk as imported from this library
+        >>> #Vdisk = #^
         >>> r = np.linspace(0, 20, 50)
         >>> diskspline = disk(r, dpref=1)
[…]
