### 2022-07-26 43a5698
Fix some bugs related to reorganization
And continue documenting
