+++ binder/python/widget_BH.py
@@ -1,3 +1,6 @@
+"""A library for handling the black hole widgets found in `10_Bonus_Black_Holes_as_DM.ipynb <https://github.com/villano-lab/galactic-spin-W1/blob/master/binder/10_Bonus_Black_Holes_as_DM.ipynb>`_.
+"""
+
 #########################
 ### Black Hole Widget ###
 #########################
@@ -112,10 +115,10 @@ def defaultnumber(galaxy):
 
 #For cutoff radius:
 minrcutBH = 0.1
-"""[float]: The minimum cutoff radius for black holes.
+"""[float]: The minimum cutoff radius for black holes, in kpc.
 """
 def maxrcutBH(galaxy):
-    """A function that returns the maximum cutoff radius for black holes appropriate for the supplied galaxy.
+    """A function that returns the maximum cutoff radius for black holes, in kpc, appropriate for the supplied galaxy.
 
     Parameters:
         galaxy : [string | int]
@@ -131,7 +134,7 @@ def maxrcutBH(galaxy):
     elif str(galaxy) == "7814" or str(galaxy).upper().replace(" ","") == 'NGC7814':
         return 4.0
 def defaultrcutBH(galaxy):
-    """A function that returns the default cutoff radius for black holes appropriate for the supplied galaxy.
+    """A function that returns the default cutoff radius for black holes, in kpc, appropriate for the supplied galaxy.
 
     Parameters:
         galaxy : [string | int]
@@ -165,7 +168,24 @@ layout = {'width':'800px'}
 
 ### NGC 5533 ###
 def f5533(arraysize,massMiniBH,rcut):
+    """
+    Generate a plot of the NGC5533 data and components alongside an image of a galaxy with dots representing black holes.
+
+    This function is intended for use as part of a widget.
+
+    Parameters:
+        arraysize: [int]
+            Size intended for the radius and angle arrays.
+        massMiniBH: [int]
+            Mass of the tiny black holes, in solar masses.
+        rcut: [float]
+            Cutoff radius for black hole placement, in kpc.
     
+    Returns: 
+        None. Generates a plot of NGC5533 data/components alongside a galaxy image.
+
+    .. seealso:: For an example usage of this function, see the notebook `10_Bonus_Black_Holes_as_DM.ipynb on Binder <https://mybinder.org/v2/gh/villano-lab/galactic-spin-W1/HEAD?labpath=binder%2F10_Bonus_Black_Holes_as_DM.ipynb>`_.
+    """
     _,fit_dict = comp.bestfit(comp.totalvelocity_miniBH,'NGC5533')
     
     bpref = fit_dict['bpref']
@@ -255,7 +275,24 @@ def f5533(arraysize,massMiniBH,rcut):
         
 ### NGC 7814 ###
 def f7814(arraysize,massMiniBH,rcut):
+    """
+    Generate a plot of the NGC7814 data and components alongside an image of a galaxy with dots representing black holes.
+
+    This function is intended for use as part of a widget.
+
+    Parameters:
+        arraysize: [int]
+            Size intended for the radius and angle arrays.
+        massMiniBH: [int]
+            Mass of the tiny black holes, in solar masses.
+        rcut: [float]
+            Cutoff radius for black hole placement, in kpc.
     
+    Returns: 
+        None. Generates a plot of NGC7814 data/components alongside a galaxy image.
+
[…]
