### 2022-07-26 6aaec6c
Set autoapi to keep autogenerated files for debug
But don't push said files.
And keep documenting.
