+++ binder/python/load_galaxies.py
@@ -38,6 +38,220 @@ NGC5533 = {
     'Mabs'             : -21.66,# B-band absolute magnitude of NGC 5533 (Norrdermeer, Van Der Hulst, 2007)
     'check'            : True
 }
+"""[dict] Data for galaxy NGC5533. Parameters and data measurements in this dictionary are from [Noordermeer2007]_ unless noted otherwise.
+
+Keys:
+    blackhole: [dict] Further information pertaining to the black hole component.
+        c: [array] B-spline coefficients returned by using scipy.interpolate.splrep on the black hole component.
+
+        k: [int] Degree of spline returned by using scipy.interpolate.splrep on the black hole component.
+        
+        Mbh: [float] Mass of the black hole.
+        
+        r: [array] Radial (kpc) values of the black hole component.
+
+        spline: [scipy.interpolate._bsplines.BSpline] Spline of the black hole component.
+        
+        t: [array] Vector of knots returned by using scipy.interpolate.splrep on the black hole component.
+
+        v: [array] Velocity (km/s) values of the black hole component.
+    
+    bulge: [dict] Further information pertaining to the bulge component.
+        c: [array] B-spline coefficients returned by using scipy.interpolate.splrep on the bulge component.
+
+        k: [int] Degree of spline returned by using scipy.interpolate.splrep on the bulge component.
+
+        Lb: [float] Luminosity (solar luminosities) of the bulge.
+
+        ML: [float] Mass-to-light ratio of the bulge.
+
+        n: [float] Concentration parameter.
+
+        q: [float] Intrinsic axis ratio.
+
+        r: [array] Radial (kpc) values of the bulge component.
+
+        re_arcsec: [float] Effective radius (arcsec).
+
+        re_rad: [float] Effective radius (radians).
+
+        spline: [scipy.interpolate._bsplines.BSpline] Spline of the bulge component.
+        
+        t: [array] Vector of knots returned by using scipy.interpolate.splrep on the bulge component.
+
+        v: [array] Velocity (km/s) values of the bulge component.
+
+    check: [bool] A flag indicating that the necessary data for the "Literature Search Check" in `07_Bonus_Bulge_Rotation_Curve.ipynb <https://github.com/villano-lab/galactic-spin-W1/blob/master/binder/07_Bonus_Bulge_Rotation_Curve.ipynb>`_ is available.
+
+    disk: [dict] Further information pertaining to the disk component.
+        c: [array] B-spline coefficients returned by using scipy.interpolate.splrep on the disk component.
+
+        k: [int] Degree of spline returned by using scipy.interpolate.splrep on the disk component.
+        
+        r: [array] Radial (kpc) values of the disk component.
+
+        spline: [scipy.interpolate._bsplines.BSpline] Spline of the disk component.
+        
+        t: [array] Vector of knots returned by using scipy.interpolate.splrep on the disk component.
+
+        v: [array] Velocity (km/s) values of the disk component.
+
+    D_Mpc: [float] The distance to the galaxy in Megaparsecs.
+
+    D_kpc: [float] The distance to the galaxy in kiloparsecs.
+
+    galaxyname: [string] Name of the galaxy.
+
+    gas: [dict] Further information pertaining to the gas component.
+        c: [array] B-spline coefficients returned by using scipy.interpolate.splrep on the gas component.
+
+        k: [int] Degree of spline returned by using scipy.interpolate.splrep on the gas component.
+        
+        r: [array] Radial (kpc) values of the gas component.
+
+        spline: [scipy.interpolate._bsplines.BSpline] Spline of the gas component.
+        
+        t: [array] Vector of knots returned by using scipy.interpolate.splrep on the gas component.
+
[…]
