### 2022-07-27 6d55dc6
Fix type formatting on attributes
