+++ binder/python/widget_SPARC.py
@@ -1,5 +1,4 @@
-"""
-A module for handling the SPARC widget, notebook `09_Widget_SPARC_Galaxies.ipynb <https://github.com/villano-lab/galactic-spin-W1/blob/master/binder/09_Widget_SPARC_Galaxies.ipynb>`_.
+"""A module for handling the SPARC widget, notebook `09_Widget_SPARC_Galaxies.ipynb <https://github.com/villano-lab/galactic-spin-W1/blob/master/binder/09_Widget_SPARC_Galaxies.ipynb>`_.
 """
 
 ################################
@@ -31,7 +30,9 @@ import components as comp
 
 # Read the file where the chosen galaxy name is located (this was modified by the user)
 textfile = open('python/chosengalaxy.txt', 'r')
-"""file: A local text file, opened in read-only mode, indicating a chosen galaxy.
+"""A local text file, opened in read-only mode, indicating a chosen galaxy.
+
+:type: file
 
 The purpose of this file is to:  
 1. Allow this chosen galaxy variable to be passed flexibly between programs.
@@ -40,7 +41,9 @@ The purpose of this file is to:
 .. seealso:: The data in this text file is stored in the :func:`galaxy <widget_SPARC.galaxy>` variable.
 """
 galaxy = textfile.read()
-"""string: The name of the chosen galaxy.
+"""The name of the chosen galaxy.
+
+:type: string
 
 .. seealso:: This string is retrieved from :func:`textfile <widget_SPARC.textfile>`.
 """
@@ -48,17 +51,23 @@ textfile.close()
 
 # Define SPARC directory
 #SPARC_file_directory='./data/sparc/'                       #note that '' means the string variable is blank
-"""string: The directory containing SPARC files.
+"""The directory containing SPARC files.
+
+:type: string
 """
 
 # Define file path for .dat files
 SPARC_file_path = './data/sparc/' + galaxy + '_rotmod.dat'
-"""string: The filename (including relative path) containing data for the :func:`chosen galaxy <galaxy>`.
+"""The filename (including relative path) containing data for the :func:`chosen galaxy <galaxy>`.
+
+:type: string
 """
 
 # Load the galaxy data
 data = np.loadtxt(SPARC_file_path)
-"""ndarray: A numpy array of data loaded from the :func:`SPARC file path <SPARC_file_path>`.
+"""A numpy array of data loaded from the :func:`SPARC file path <SPARC_file_path>`.
+
+:type: array
 """
 
 # Split columns into arrays
@@ -67,21 +76,27 @@ Rad,Vobs,errV,Vgas,Vdisk,Vbul,SBdisk,SBbul = data.T
 # Check if it has all components
 if np.sum(Vbul) == 0:     # Bulge
     warning_bulge = "There is no bulge component."
-    """string: If the bulge component is missing, a message indicating that it is. Otherwise, an empty string.
+    """If the bulge component is missing, a message indicating that it is. Otherwise, an empty string.
+
+    :type: string
     """
 else: 
     warning_bulge = ""
 
 if np.sum(Vdisk) == 0:    # Disk
     warning_disk = "There is no disk component."
-    """string: If the disk component is missing, a message indicating that it is. Otherwise, an empty string.
+    """If the disk component is missing, a message indicating that it is. Otherwise, an empty string.
+
+    :type: string
     """
 else: 
     warning_disk = ""
 
 if np.sum(Vgas) == 0:     # Gas
     warning_gas = "There is no gas component."
[…]
