### 2022-07-27 2784cb5
Write some examples for components.py
And ignore dummy hdf5 file used in said examples.
