+++ binder/python/components.py
@@ -32,6 +32,10 @@ except ModuleNotFoundError:
     print("Could not find h5py. Datasets will not be able to be saved or loaded using components.py.")
 
 defaultpath = '../'
+"""The default location of files containing cached calculations.
+
+:type: string
+"""
 
 #===============================
 #========= Constants ===========
@@ -55,11 +59,14 @@ def galdict(galaxy):
         >>> # Define a function that prints the cutoff radius of any galaxy and returns it
         >>> def rcut(galaxy):
         >>>     galaxydata = galdict(galaxy) # Retrieve the whole dictionary
-        >>>     cutoff = galaxydata["rcut"]
+        >>>     cutoff = galaxydata["rc"]
         >>>     print(cutoff)
         >>>     return cutoff
         >>> # Print and assign the cutoff for NGC 5533
         >>> rcut5533 = rcut('NGC5533')
+        1.4
+        >>> print(rcut5533)
+        1.4
     """
     
     return globals()[galaxy.upper().replace(" ","")]        
@@ -68,27 +75,77 @@ def galdict(galaxy):
 
 #---------Definitely Constant---------
 G = 4.30091e-6                    # Gravitational constant (kpc/solar mass*(km/s)^2) 
-"""double: Gravitational constant in kpc/(solar mass * (km/s)^2)
+"""Gravitational constant in kpc/(solar mass * (km/s)^2)
+
+:type: double
 """
 
 #---------Measured Indirectly---------
 ups = 2.8                         # Bulge mass-to-light ratio (Solar Mass/Solar Luminosity). Source: Noordermeer, 2008
+"""Bulge mass-to-light ratio (Solar masses/Solar luminosities). [Noordermeer2008]_
+
+:type: float
+"""
 q = 0.33                          # Intrinsic axis ratio. Source: Noordermeer, 2008
+"""Intrinsic axis ratio. [Noordermeer2008]_
+
+:type: float
+"""
 e2 = 1-(q**2)                     # Eccentricity. Source: Noordermeer, 2008
+"""Bulge eccentricity. [Noordermeer2008]_
+
+:type: float
+"""
 i = 52*(np.pi/180)                # Inclination angle. Source: Noordermeer & Van Der Hulst, 2007
+"""Inclination angle (radians). [Noordermeer2008]_
+
+:type: float
+"""
 h_rc = 1.4                        # Core radius (kpc). Source: Noordermeer, 2008
+"""Core radius (kpc). [Noordermeer2008]_
+
+:type: float
+"""
 Mbh_def = 2.7e9                   # Black Hole mass (in solar mass). Source: Noordermeer, 2008
+"""Central black hole mass (Solar masses). [Noordermeer2008]_
+
+:type: float
+"""
 
 #---------Definitely Variable---------
 n_c = 2.7                         # Concentration parameter. Source: Noordermeer & Van Der Hulst, 2007
+"""Concentration parameter. [Noordermeer2007]_
+
+:type: float
+"""
 h_c = 8.9                         # Radial scale-length (kpc). Source: Noordermeer & Van Der Hulst, 2007
+"""Radial scale length (kpc). [Noordermeer2007]_
+
+:type: float
+"""
 hrho00_c = 0.31e9                 # Halo central surface density (solar mass/kpc^2). Source: Noordermeer, 2008
+"""Central surface density of halo (Solar masses/kpc^2). [Noordermeer2008]_
+
+:type: float
+"""
 drho00_c = 0.31e9                 # Disk central surface density (solar mass/kpc^2)
[…]
