### 2022-07-27 b5edb9a
[JB] added citation to other galaxies, removed M31 data and Milky Way data (not sure if we use this anywhere)
