+++ binder/python/load_galaxies.py
@@ -893,7 +893,7 @@ NGC5005['massbh'] = 0                # central black hole is included in the bul
 
 # NGC 3198
 NGC3198 = {'measured_data' : dp.getXYdata_wYerr('data/othergalaxies/NGC3198.txt')}
-"""Data for galaxy NGC3198. Parameters and data measurements in this dictionary are from [SPARC2016]_ unless noted otherwise.
+"""Data for galaxy NGC3198. Parameters and data measurements in this dictionary are from [Karukes2015]_ unless noted otherwise.
 
 Keys:
     galaxyname: [string] Name of the galaxy.
@@ -920,7 +920,7 @@ NGC3198['galaxyname'] = 'NGC 3198'
 
 # UGC 477
 UGC0477 = {'measured_data' : dp.getXYdata_wYerr('data/othergalaxies/UGC477.txt')}
-"""Data for galaxy NGC477. Parameters and data measurements in this dictionary are from [SPARC2016]_ unless noted otherwise.
+"""Data for galaxy UGC477. Parameters and data measurements in this dictionary are from [deNaray2006]_ and [deNaray2008]_ unless noted otherwise.
 
 Keys:
     galaxyname: [string] Name of the galaxy.
@@ -949,7 +949,7 @@ UGC477 = UGC0477
 
 # UGC 1281
 UGC1281 = {'measured_data' : dp.getXYdata_wYerr('data/othergalaxies/UGC1281.txt')}
-"""Data for galaxy NGC1281. Parameters and data measurements in this dictionary are from [SPARC2016]_ unless noted otherwise.
+"""Data for galaxy UGC1281. Parameters and data measurements in this dictionary are from [deNaray2006]_ and [deNaray2008]_ unless noted otherwise.
 
 Keys:
     galaxyname: [string] Name of the galaxy.
@@ -975,7 +975,7 @@ UGC1281['galaxyname'] = 'UGC 1281'
 
 # UGC 1437
 UGC1437 = {'measured_data' : dp.getXYdata_wYerr('data/othergalaxies/UGC1437.txt')}
-"""Data for galaxy NGC1437. Parameters and data measurements in this dictionary are from [SPARC2016]_ unless noted otherwise.
+"""Data for galaxy UGC1437. Parameters and data measurements in this dictionary are from [Epinat2008]_ unless noted otherwise.
 
 Keys:
     galaxyname: [string] Name of the galaxy.
@@ -1001,7 +1001,7 @@ UGC1437['galaxyname'] = 'UGC 1437'
 
 # UGC 2953
 UGC2953 = {'measured_data' : dp.getXYdata_wYerr('data/othergalaxies/UGC2953.txt')}
-"""Data for galaxy NGC2953. Parameters and data measurements in this dictionary are from [SPARC2016]_ unless noted otherwise.
+"""Data for galaxy UGC2953. Parameters and data measurements in this dictionary are from [SPARC2016]_ unless noted otherwise.
 
 Keys:
     galaxyname: [string] Name of the galaxy.
@@ -1027,7 +1027,7 @@ UGC2953['galaxyname'] = 'UGC 2953'
 
 # UGC 4325
 UGC4325 = {'measured_data' : dp.getXYdata_wYerr('data/othergalaxies/UGC4325.txt')}
-"""Data for galaxy NGC4325. Parameters and data measurements in this dictionary are from [SPARC2016]_ unless noted otherwise.
+"""Data for galaxy UGC4325. Parameters and data measurements in this dictionary are from [deNaray2006]_ and [deNaray2008]_ unless noted otherwise.
 
 Keys:
     galaxyname: [string] Name of the galaxy.
[…]
