### 2022-07-27 b61a3ea
[JB] added example to set_params and description to h5py (feel free to edit)
