### 2022-07-27 184b995
[JB] added type for widget buttons
