### 2022-07-27 367b188
[JB] changed galaxy name, description for arraysize
