+++ binder/python/widget_BH.py
@@ -84,12 +84,14 @@ stepN = 5
 
 :int:
 """
+
 def minnumberBH(galaxy):
     """A function that returns the minimum number of black holes (prior to multiplying by the :func:`scale <widget_BH.scale>`) appropriate for the supplied galaxy.
 
     Parameters:
         galaxy : [string | int]
-            The name or number (for NGC) of the selected galaxy. Names are not case-sensitive and ignore spaces. Allowed inputs: "NGC5533" or "NGC7814".
+            The name or number (for NGC) of the selected galaxy. Names are not case-sensitive and ignore spaces. 
+            Allowed inputs: "NGC5533" or "NGC7814".
 
     Returns: [int] 
         The minimum number of black holes, prior to multiplying by the :func:`scale <widget_BH.scale>`.
@@ -100,12 +102,14 @@ def minnumberBH(galaxy):
         return 50
     elif str(galaxy) == "7814" or str(galaxy).upper().replace(" ","") == 'NGC7814':
         return 1
+    
 def maxnumberBH(galaxy):
     """A function that returns the maximum number of black holes (prior to multiplying by the :func:`scale <widget_BH.scale>`) appropriate for the supplied galaxy.
 
     Parameters:
         galaxy : [string | int]
-            The name or number (for NGC) of the selected galaxy. Names are not case-sensitive and ignore spaces. Allowed inputs: "NGC5533" or "NGC7814".
+            The name or number (for NGC) of the selected galaxy. Names are not case-sensitive and ignore spaces. 
+            Allowed inputs: "NGC5533" or "NGC7814".
 
     Returns: 
         [int] The maximum number of black holes, prior to multiplying by the :func:`scale <widget_BH.scale>`.
@@ -116,12 +120,14 @@ def maxnumberBH(galaxy):
         return 5e2
     elif str(galaxy) == "7814" or str(galaxy).upper().replace(" ","") == 'NGC7814':
         return 1000
+    
 def defaultnumber(galaxy):
     """A function that returns the default number of black holes (prior to multiplying by the :func:`scale <widget_BH.scale>`) appropriate for the supplied galaxy.
 
     Parameters:
         galaxy : [string | int]
-            The name or number (for NGC) of the selected galaxy. Names are not case-sensitive and ignore spaces. Allowed inputs: "NGC5533" or "NGC7814".
+            The name or number (for NGC) of the selected galaxy. Names are not case-sensitive and ignore spaces. 
+            Allowed inputs: "NGC5533" or "NGC7814".
 
     Returns: 
         [int] The default number of black holes, prior to multiplying by the :func:`scale <widget_BH.scale>`.
@@ -139,12 +145,14 @@ minrcutBH = 0.1
 
 :type: float
 """
+
 def maxrcutBH(galaxy):
     """A function that returns the maximum cutoff radius for black holes, in kpc, appropriate for the supplied galaxy.
 
     Parameters:
         galaxy : [string | int]
-            The name or number (for NGC) of the selected galaxy. Names are not case-sensitive and ignore spaces. Allowed inputs: "NGC5533" or "NGC7814".
+            The name or number (for NGC) of the selected galaxy. Names are not case-sensitive and ignore spaces. 
+            Allowed inputs: "NGC5533" or "NGC7814".
 
     Returns: 
         [float] The maximum cutoff radius for black holes.
@@ -155,12 +163,14 @@ def maxrcutBH(galaxy):
         return 3.0
     elif str(galaxy) == "7814" or str(galaxy).upper().replace(" ","") == 'NGC7814':
         return 4.0
+    
 def defaultrcutBH(galaxy):
[…]
