### 2022-07-27 52cc030
[JB] removed >>> from example output and some not needed comments
