### 2022-07-27 25595e5
Adjust template to work for load_galaxies
