+++ binder/python/components.py
@@ -27,7 +27,7 @@ from load_galaxies import *
 try:
     import h5py as h5
     h5py = 1
-    """If h5py = 1, there is data to be loaded.
+    """If `1`, the h5py library is available for saving and loading data. If `0`, then all calculations must be done each time without saving or loading.
 
     :type: int
     """
@@ -49,17 +49,17 @@ def galdict(galaxy):
     """
     Retrieve a dictionary of parameters for the associated galaxy.
 
-    Parameters:
+    :parameters:
         galaxy: [string]
             The galaxy's full name, including catalog. Not case-sensitive. Ignores spaces. 
 
-    Returns:
+    :returns:
         Dictionary of parameters for the associated galaxy.
 
     .. note::
-        For information on the parameters contained in the returned dictionary, see the documentation for `load_galaxies.py`.
+        For information on the parameters contained in the returned dictionary, see the documentation for `load_galaxies.py <../load_galaxies/index.html>`_.
 
-    Example:
+    :example:
         >>> # Define a function that prints the cutoff radius of any galaxy and returns it
         >>> def rcut(galaxy):
         >>>     galaxydata = galdict(galaxy) # Retrieve the whole dictionary
@@ -79,14 +79,14 @@ def galdict(galaxy):
 
 #---------Definitely Constant---------
 G = 4.30091e-6                    # Gravitational constant (kpc/solar mass*(km/s)^2) 
-"""Gravitational constant in kpc/(solar mass * (km/s)^2)
+"""Gravitational constant (:math:`kpc/(M_{Sun}(km/s)^2)`).
 
 :type: double
 """
 
 #---------Measured Indirectly---------
 ups = 2.8                         # Bulge mass-to-light ratio (Solar Mass/Solar Luminosity). Source: Noordermeer, 2008
-"""Bulge mass-to-light ratio (Solar masses/Solar luminosities). [Noordermeer2008]_
+"""Bulge mass-to-light ratio (:math:`M_{Sun}/L_{Sun}`). [Noordermeer2008]_
 
 :type: float
 """
@@ -106,12 +106,12 @@ i = 52*(np.pi/180)                # Inclination angle. Source: Noordermeer & Van
 :type: float
 """
 h_rc = 1.4                        # Core radius (kpc). Source: Noordermeer, 2008
-"""Core radius (kpc). [Noordermeer2008]_
+"""Core radius (:math:`kpc`). [Noordermeer2008]_
 
 :type: float
 """
 Mbh_def = 2.7e9                   # Black Hole mass (in solar mass). Source: Noordermeer, 2008
-"""Central black hole mass (Solar masses). [Noordermeer2008]_
+"""Central black hole mass (:math:`M_{Sun}`). [Noordermeer2008]_
 
 :type: float
 """
@@ -123,24 +123,24 @@ n_c = 2.7                         # Concentration parameter. Source: Noordermeer
 :type: float
 """
 h_c = 8.9                         # Radial scale-length (kpc). Source: Noordermeer & Van Der Hulst, 2007
-"""Radial scale length (kpc). [Noordermeer2007]_
+"""Radial scale length (:math:`kpc`). [Noordermeer2007]_
 
 :type: float
 """
 hrho00_c = 0.31e9                 # Halo central surface density (solar mass/kpc^2). Source: Noordermeer, 2008
-"""Central surface density of halo (Solar masses/kpc^2). [Noordermeer2008]_
+"""Central surface density of halo (:math:`M_{Sun}/kpc^2`). [Noordermeer2008]_
 
 :type: float
 """
 drho00_c = 0.31e9                 # Disk central surface density (solar mass/kpc^2)
[…]
