+++ binder/python/widget_BH.py
@@ -36,12 +36,12 @@ center = [1960,1800]
 
 # Kpc limits, visual guess based on the galaxy in the image chosen:
 minkpc = 0
-"""The minimum value, in kpc, allowed for the radial position of black holes on the plot.
+"""The minimum value (:math:`kpc`), allowed for the radial position of black holes on the plot.
 
 :type: int
 """
 maxkpc = 100
-"""The maximum value, in kpc, allowed for the radial position of black holes on the plot.
+"""The maximum value (:math:`kpc`), allowed for the radial position of black holes on the plot.
 
 :type: int
 """
@@ -88,12 +88,12 @@ stepN = 5
 def minnumberBH(galaxy):
     """A function that returns the minimum number of black holes (prior to multiplying by the :func:`scale <widget_BH.scale>`) appropriate for the supplied galaxy.
 
-    Parameters:
+    :parameters:
         galaxy : [string | int]
             The name or number (for NGC) of the selected galaxy. Names are not case-sensitive and ignore spaces. 
             Allowed inputs: "NGC5533" or "NGC7814".
 
-    Returns: [int] 
+    :returns: [int] 
         The minimum number of black holes, prior to multiplying by the :func:`scale <widget_BH.scale>`.
     
     .. seealso:: For an example usecase of this function, see :func:`arraysize_5533 <widget_BH.arraysize_5533>`.
@@ -106,12 +106,12 @@ def minnumberBH(galaxy):
 def maxnumberBH(galaxy):
     """A function that returns the maximum number of black holes (prior to multiplying by the :func:`scale <widget_BH.scale>`) appropriate for the supplied galaxy.
 
-    Parameters:
+    :parameters:
         galaxy : [string | int]
             The name or number (for NGC) of the selected galaxy. Names are not case-sensitive and ignore spaces. 
             Allowed inputs: "NGC5533" or "NGC7814".
 
-    Returns: 
+    :returns: 
         [int] The maximum number of black holes, prior to multiplying by the :func:`scale <widget_BH.scale>`.
     
     .. seealso:: For an example usecase of this function, see :func:`arraysize_5533 <widget_BH.arraysize_5533>`.
@@ -124,12 +124,12 @@ def maxnumberBH(galaxy):
 def defaultnumber(galaxy):
     """A function that returns the default number of black holes (prior to multiplying by the :func:`scale <widget_BH.scale>`) appropriate for the supplied galaxy.
 
-    Parameters:
+    :parameters:
         galaxy : [string | int]
             The name or number (for NGC) of the selected galaxy. Names are not case-sensitive and ignore spaces. 
             Allowed inputs: "NGC5533" or "NGC7814".
 
-    Returns: 
+    :returns: 
         [int] The default number of black holes, prior to multiplying by the :func:`scale <widget_BH.scale>`.
     
     .. seealso:: For an example usecase of this function, see :func:`arraysize_5533 <widget_BH.arraysize_5533>`.
@@ -141,20 +141,20 @@ def defaultnumber(galaxy):
 
 #For cutoff radius:
 minrcutBH = 0.1
-"""The minimum cutoff radius for black holes, in kpc.
+"""The minimum cutoff radius for black holes (:math:`kpc`).
 
 :type: float
 """
 
 def maxrcutBH(galaxy):
-    """A function that returns the maximum cutoff radius for black holes, in kpc, appropriate for the supplied galaxy.
+    """A function that returns the maximum cutoff radius for black holes (:math:`kpc`), appropriate for the supplied galaxy.
 
-    Parameters:
+    :parameters:
         galaxy : [string | int]
[…]
