+++ docs/source/conf.py
@@ -18,10 +18,6 @@ import subprocess, os, sys
 sys.path.append('../binder/python')
 #real quick, make a link if it doesn't exist
 os.chdir('../')
-try:
-    os.symlink('../binder/data','data')
-except FileExistsError:
-    pass
 os.chdir('source')
 
 autodoc_mock_imports = ['matplotlib','IPython','astroquery','astropy','ipywidgets']

[…]
