### 2022-07-29 #177 @villaa
@jgorzalski My team has made some upgrades so we would like to change the review version here to [v3.0.0](https://github.com/villano-lab/galactic-spin-W1/releases/tag/v3.0.0), thanks.
