### 2022-08-06 #177 @villaa
@labarba I don't know if you're able, but please update the review version here to [v3.0.0](https://github.com/villano-lab/galactic-spin-W1/releases/tag/v3.0.0), thanks.
