### 2022-09-06 #177 @moorepants
@IanHawke, @ali-ramadhan, @ntmoore, and @Carreau would any of you like to review this JOSE submission about an education module for simulating galaxies?
