### 2022-09-13 #184 @whedon
[REVIEW]: The Data Behind Dark Matter: Exploring Galactic Rotation
**Submitting author:** @villaa (<a href="http://orcid.org/0000-0002-3893-7259">Anthony Villano</a>)
**Repository:** https://github.com/villano-lab/galactic-spin-W1
**Branch with paper.md** (empty if default branch):
**Version:** v3.0.4
