### 2022-09-13 #184 @moorepants
The missing DOI leads to a login page:
![image](https://user-images.githubusercontent.com/276007/189932232-4e6700eb-8c87-4311-ac0c-05d3e35ba62d.png)
Is there a way to link to a public version of the paper?
