### 2022-09-13 #184 @nuclearGoblin
That DOI links to a book that does not appear to have a public copy, but it does have [a page on the AIP website](https://uscibooks.aip.org/books/introduction-to-error-analysis-2nd-ed/), would that be an appropriate link?
