### 2022-09-13 71fabfb
Addressing pre-review comments Missing space, comments on binder Also added a link for book whose DOI is behind a login.
