+++ paper/references.bib
@@ -193,7 +193,8 @@
   isbn = {093570275X},
   publisher = {University Science Books},
   title = {An Introduction to Error Analysis: The Study of Uncertainties in Physical Measurements},
-  year = 1996
+  year = 1996,
+  url = {https://uscibooks.aip.org/books/introduction-to-error-analysis-2nd-ed/}
 }
 
 @misc{ScientificAmerican2008, 
@@ -366,3 +367,13 @@ archivePrefix = {arXiv},
   url = {https://www.ucdenver.edu/sites/research-day/event-details#ac-special-session-2-dark-matter-workshop-0},
   publisher = {The University of Colorado Denver},
   urldate={06-06-2022}}
+
+@InProceedings{ Binder,
+  author    = { {P}roject {J}upyter and {M}atthias {B}ussonnier and {J}essica {F}orde and {J}eremy {F}reeman and {B}rian {G}ranger and {T}im {H}ead and {C}hris {H}oldgraf and {K}yle {K}elley and {G}ladys {N}alvarte and {A}ndrew {O}sheroff and {M} {P}acer and {Y}uvi {P}anda and {F}ernando {P}erez and {B}enjamin {R}agan-{K}elley and {C}arol {W}illing },
+  title     = { {B}inder 2.0 - {R}eproducible, interactive, sharable environments for science at scale },
+  booktitle = { {P}roceedings of the 17th {P}ython in {S}cience {C}onference },
+  pages     = { 113 - 120 },
+  year      = { 2018 },
+  editor    = { {F}atih {A}kici and {D}avid {L}ippa and {D}illon {N}iederhut and {M} {P}acer },
+  doi       = { 10.25080/Majora-4af1f417-011 }
+}
