### 2022-09-13 #184 @moorepants
I would just cite the book as a traditional citation sans the DOI, as the DOI doesn't seem to lead to the book.
