+++ paper/references.bib
@@ -193,8 +193,7 @@
   isbn = {093570275X},
   publisher = {University Science Books},
   title = {An Introduction to Error Analysis: The Study of Uncertainties in Physical Measurements},
-  year = 1996,
-  url = {https://uscibooks.aip.org/books/introduction-to-error-analysis-2nd-ed/}
+  year = 1996
 }
 
 @misc{ScientificAmerican2008, 
