### 2022-09-14 ad1253b
Update URL
Also do you mean `latest`, or `stable` ? Latest is usually for dev version and stable for most recent tag.
