### 2022-09-14 #184 @Carreau
Linking to [comments in pre-review](https://github.com/openjournals/jose-reviews/issues/177#issuecomment-1239057732):

> ```
> Sérsic bulges(Edo Noordermeer,2008),
> ```
> Missing space before the opening bracket if we compare to other citations – I don't know if the authors want to fix that.

> I also see on the repository that the author have packaged everything to be available on Binder – that I have started to play with and would have love to have 15 years ago in my cosmo classes – maybe the author want to say one or two sentence as to whether they feel it is better to have local install for users to retain their work and. have a working environment, or prefer the ease of use of the an ephemeral binder ?

I finished going through the notebooks and material, and it is a really great content, that show some limitation of Jupyter like having to guide users to "Run all", and sometime the lack of obvious indication that computation is happening, but that is something the jupyter team should fix not the author. I found 1 broken link, for which I sent a PRs.
