### 2022-09-14 1080fce
Try supplying matplotlib version to fix discrepancies in plot sizes
