### 2022-09-14 9ed5b0c
Specify h5py version to prevent incompatibilities with saved data
