### 2022-09-15 be82f4a
Try specifying jupyterlab version
