### 2022-09-15 7659e1e
Specify ipympl version
