### 2022-09-15 3262e03
Specify ipylab version Still not sure why there is an issue with fig size as that should be controlled by the html function specifying it?
