+++ binder/environment.yml
@@ -11,7 +11,7 @@ dependencies:
   - h5py==3.6.0
   - sympy
   - ipympl==0.8.7
-  - ipylab
+  - ipylab==0.5.2
   - ipywidgets==7.6.5
   - ffmpeg
   - pip
