### 2022-09-16 #184 @Carreau
> I found 1 broken link, for which I sent a PRs.

https://github.com/villano-lab/galactic-spin-W1/pull/32

I actually forgot the press the "Send PR" button...
