### 2022-09-20 f36fbc8
Specify python and nbval versions Maybe the issue is with the test itself
