+++ .travis.yml
@@ -1,5 +1,5 @@
 language: python
-python: 3.7
+python: 3.7.9
 dist: xenial
 sudo: true
 install:
@@ -22,7 +22,7 @@ install:
   - conda env create -n test-environment python=$TRAVIS_PYTHON_VERSION -f binder/environment.yml
   - source activate test-environment
   #Packages for testing only -- don't need in user environment.
-  - pip3 install codecov pytest-cov pytest_notebook nbval papermill
+  - pip3 install codecov pytest==7.1.3 pytest-cov pytest_notebook nbval==0.9.6 papermill
   - ipython kernel install --user --name=galactic-spin 
 
 before_script: cd binder
