### 2022-09-20 ba9be4d
Specify figure sizes in pixels to get consistency in pytest across displays.
