### 2022-09-21 3d8fd2a
Change links to 'latest' docs to point to stable where appropriate
