+++ README.md
@@ -13,8 +13,8 @@
 The first workshop of our Rotation Curves series, which covers the basics of what rotation curves are,
 how they are measured, and how we model them by breaking them into components.
 
-You can find more detailed documentation about the code involved in this workshop [here](https://galactic-spin-W1.readthedocs.io/en/latest), 
-including [a page guiding you through the workshop itself](https://galactic-spin-w1.readthedocs.io/en/latest/01_Getting_Started.html).
+You can find more detailed documentation about the code involved in this workshop [here](https://galactic-spin-W1.readthedocs.io/en/stable), 
+including [a page guiding you through the workshop itself](https://galactic-spin-w1.readthedocs.io/en/stable/01_Getting_Started.html).
 
 ## ABSTRACT
 
