### 2022-10-26 #184 @villaa
> @villaa Might you have more suggestions for reviewers? I'm not from this field or research/education. Maybe you know more?

I don't know anyone else on the reviewer list; but I might know some others who have the skills to do it.

I'm not sure if @matthewfeickert would be able to review, but they are very well qualified in all things code and strong overlap with my type of science.

I'll look for others in my field who might be good; shall I send you their emails separately?
