### 2022-10-26 #184 @matthewfeickert
Hi @villaa! :wave: Thanks for thinking of me. I'm unfortunately a bit over committed with work until early December, so I think for the timeline you need I won't be of much help &mdash; sorry!

I just wanted to say though that I think that your workshop looks excellent and I think it is fantastic that you're submitting to JOSE! Congratulations in advance on the great work! :rocket:
