### 2022-11-08 #184 @nuclearGoblin
> @\villaa Might you have more suggestions for reviewers? I'm not from this field or research/education. Maybe you know more?

(I hope it's okay for me to give some suggestions as well, I can also send a list of a few reviewers if you like.) From the JOSS reviewers list, would @\dpshelio be able to review? It looks like he is familiar with physics programming, python, and astrophysics.
