### 2022-12-14 #184 @moorepants
Hi @dpshelio, might you be interested in reviewing this Journal of Open Source Education paper on "The Data Behind Dark Matter: Exploring Galactic Rotation"?
