### 2022-12-20 #184 @villaa
@mattbellis are you willing to review this submission for JOSE--it's a workshop/tutorial on dark matter physics. We've struggled to find a second reviewer so it might even be OK if you can only get to it with a significant delay (say sometime in the new year).

Anyway, it occurred to me you'd be an excellent reviewer here. If you can't, no worries, hope all is well.
