### 2023-01-10 #184 @mattbellis
Sure. I'm an associate professor in the department of physics and astronomy at Siena College. I've been here since 2012.

https://www.siena.edu/faculty-and-staff/person/matthew-bellis/

My primary research is particle physics. I'm currently a member of the CMS and BaBar collaborations and prior to that, I worked with the CLAS and GlueX experiments at Jefferson Lab. I've also collaborated with others and presented posters on work with open data released by the CoGeNT dark matter experiment. Part of my work with CMS involved new dark matter signatures.

I've published work on using GPUs to speed up calculations relevant to large-scale structure in the universe.

I've done a fair amount of work within Siena related to computing in the classroom and I maintain a website with students, Particle Physics Playground that provides simplified data and tutorials for anyone to use.

https://particle-physics-playground.github.io/

I've published in JOSS previously, though never served as a reviewer.

https://joss.theoj.org/papers/10.21105/joss.00384
