### 2023-01-10 #184 @moorepants
Ok, it looks correct for me now. @mattbellis make sure to read the https://joss.readthedocs.io/en/latest/reviewer_guidelines.html and let me know if you have any questions. Thanks for volunteering.
