### 2023-02-06 #184 @labarba
Hi @villaa — could I ask you to address your queries to the handling editor? As you know, everyone is a volunteer here, and we find that being considerate to the different roles is a good idea. Your handling editor also knows if they have have some other thread going with the reviewers (via email or other comms). And thank you for your patience with the review process! 🙏
