### 2023-04-06 #184 @villaa
> Hi @villaa, I see that @besk467 was at your same institution. Are there any conflict of interests to declare? If not, they look like a good fit to do this review.

@moorepants we were actually never at the same institution. @besk467 was at the University of Denver (DU) and I am at the University of Colorado Denver. Different schools.

We do know each other and have met outside of work a handful of times, but if @besk467 does not think it will bias the review, I am inclined to agree.
