### 2023-04-07 8cae139
Remove more calls to cached values.
