+++ .circleci/config.yml
@@ -74,33 +74,33 @@ jobs:
             echo "1. Preparing building environment"
             echo "Working directory is ${PWD}"
             # If not set, define DOCKER_TAG
-            #if [ ! -n "${DOCKER_TAG:-}" ]
-            #    then
-            DOCKER_TAG=${CIRCLE_SHA1}
-            #fi
+            if [ ! -n "${DOCKER_TAG:-}" ]
+                then
+                    DOCKER_TAG=${CIRCLE_SHA1}
+            fi
             # If not set, define CONTAINER_NAME
-            #if [ ! -n "${CONTAINER_NAME:-}" ]
-            #    then
-            CONTAINER_NAME="${CIRCLE_PROJECT_USERNAME}/${CIRCLE_PROJECT_REPONAME}"
-            #fi
+            if [ ! -n "${CONTAINER_NAME:-}" ]
+                then
+                    CONTAINER_NAME="${CIRCLE_PROJECT_USERNAME}/${CIRCLE_PROJECT_REPONAME}"
+            fi
             echo "Container name is ${CONTAINER_NAME}"
             # If not set, define REPO_NAME
-            #if [ ! -n "${REPO_NAME:-}" ]
-            #    then
-            LOCAL_REPO="${CIRCLE_PROJECT_USERNAME}/${CIRCLE_PROJECT_REPONAME}"
-            REPO_NAME=/tmp/src
-            echo "Repository name (REPO_NAME) is not defined, will build ${LOCAL_REPO} in ${REPO_NAME}"
-            ls ${REPO_NAME}
-            #else
-            #    echo "Repository name found defined for build: ${REPO_NAME}"
-            #fi
+            if [ ! -n "${REPO_NAME:-}" ]
+                then
+                    LOCAL_REPO="${CIRCLE_PROJECT_USERNAME}/${CIRCLE_PROJECT_REPONAME}"
+                    REPO_NAME=/tmp/src
+                    echo "Repository name (REPO_NAME) is not defined, will build ${LOCAL_REPO} in ${REPO_NAME}"
+                    ls ${REPO_NAME}
+            else
+                echo "Repository name found defined for build: ${REPO_NAME}"
+            fi
             echo "2. Running jupyter-repo2docker..."
             set -x
-            #if docker pull "${CONTAINER_NAME}:${DOCKER_TAG}"; then
-            #  jupyter-repo2docker --debug --user-name jovyan --user-id 1000 --no-run --image-name "${CONTAINER_NAME}:${DOCKER_TAG}" --cache-from "${CONTAINER_NAME}:${DOCKER_TAG}" "${REPO_NAME}"
-            #else
-            jupyter-repo2docker --debug --user-name jovyan --user-id 1000 --no-run --image-name "${CONTAINER_NAME}:${DOCKER_TAG}" "${REPO_NAME}"
-            #fi
+            if docker pull "${CONTAINER_NAME}:${DOCKER_TAG}"; then
+              jupyter-repo2docker --debug --user-name jovyan --ref "{CIRCLE_SHA1}" --user-id 1000 --no-run --image-name "${CONTAINER_NAME}:${DOCKER_TAG}" --cache-from "${CONTAINER_NAME}:${DOCKER_TAG}" "${REPO_NAME}"
+            else
+              jupyter-repo2docker --debug --user-name jovyan --ref "{CIRCLE_SHA1}" --user-id 1000 --no-run --image-name "${CONTAINER_NAME}:${DOCKER_TAG}" "${REPO_NAME}"
+            fi
             docker ps
             docker images
       - run:
