+++ .circleci/config.yml
@@ -97,9 +97,9 @@ jobs:
             echo "2. Running jupyter-repo2docker..."
             set -x
             if docker pull "${CONTAINER_NAME}:${DOCKER_TAG}"; then
-              jupyter-repo2docker --debug --user-name jovyan --ref "{CIRCLE_SHA1}" --user-id 1000 --no-run --image-name "${CONTAINER_NAME}:${DOCKER_TAG}" --cache-from "${CONTAINER_NAME}:${DOCKER_TAG}" "${REPO_NAME}"
+              jupyter-repo2docker --debug --user-name jovyan --ref "${CIRCLE_SHA1}" --user-id 1000 --no-run --image-name "${CONTAINER_NAME}:${DOCKER_TAG}" --cache-from "${CONTAINER_NAME}:${DOCKER_TAG}" "${REPO_NAME}"
             else
-              jupyter-repo2docker --debug --user-name jovyan --ref "{CIRCLE_SHA1}" --user-id 1000 --no-run --image-name "${CONTAINER_NAME}:${DOCKER_TAG}" "${REPO_NAME}"
+              jupyter-repo2docker --debug --user-name jovyan --ref "${CIRCLE_SHA1}" --user-id 1000 --no-run --image-name "${CONTAINER_NAME}:${DOCKER_TAG}" "${REPO_NAME}"
             fi
             docker ps
             docker images
