+++ RELEASENOTES.md
@@ -1,3 +1,8 @@
+## Release (v3.0.1) Date 07.04.23
+Hotfix for restoring Binder functionality.
+* The repository now successfully builds on Binder again.
+* Fixed CircleCI builds only retrieving files from master for testing; they now retrieve files from the commit being tested.
+
 ## Release (v3.0.0) Date 28.07.22
 
 Create documentation oriented toward developers and reorganize libraries.
@@ -9,7 +14,6 @@ Create documentation oriented toward developers and reorganize libraries.
 
 * Improved notebook 06 by restoring the ability to "Run all cells below" with a button press instead of JupyterLab UI.
 * Added another CI to test an environment generated more similarly to how Binder generates its environments.
-
 
 ## Release (v2.0.0) Date 24.06.22
 
