+++ binder/environment.yml
@@ -3,7 +3,7 @@ channels:
   - conda-forge
   - defaults
 dependencies:
-  - python=3.7.9
+  - python==3.7.9
   - numpy
   - scipy
   - matplotlib
