### 2023-04-08 d1d5be6
Specify DPI for consistent plot sizing for tests.
