+++ binder/python/widget_SPARC.py
@@ -564,7 +564,7 @@ def GalaxyImage(position=galaxy,survey=['DSS']):
     gfilter = hdu.data
 
     # WCS 
-    plt.figure(figsize=(5,5))
+    plt.figure(figsize=(5,5),dpi=100)
     wcs = WCS(hdu.header)
     ax = plt.gca(projection=wcs)
 
