### 2023-04-17 #184 @besk467
@moorepants  Hi. I agree, no conflict of interest and I'd like to review Anthony's paper. Thanks. my email is sofcis94@uw.edu
I'm at University of Washington now.
