### 2023-04-25 #184 @moorepants
It was supposed to do that when we swapped reviewers, but it didn't seem to. I don't know how to fix that other than manually editing it (if that is possible).

@besk467 did you accept this invitation; https://github.com/openjournals/jose-reviews/invitations? Maybe that triggers the change of the name.
