### 2023-04-25 #184 @labarba
hi @moorepants — the editorial bot is not able to make edits to the issue comments. Unfortunately, you do have to edit the comment manually to add a checklist for a new reviewer (or change the named reviewer of an existing checklist).
