### 2023-05-08 #184 @moorepants
Yes, you should provide a general statement about your opinion on the paper and software. If you find any issues that should be addressed before accepting for publishing, then you should note them here in your comments or open issues on the repository.

Here are some good examples:

- https://github.com/openjournals/jose-reviews/issues/44#issuecomment-459029461
- https://github.com/openjournals/jose-reviews/issues/115#issuecomment-840915670
- https://github.com/openjournals/jose-reviews/issues/148#issuecomment-976835904
