### 2023-05-10 #184 @moorepants
@besk467, thank you that is perfect. And very nice to see all is in order and you have a positive review.

@Carreau, would you mind providing an overall statement for your review also? https://github.com/openjournals/jose-reviews/issues/184#issuecomment-1246474264 does not make a recommendation for publishing or whether all issues have been addressed.
