### 2023-05-17 #184 @moorepants
I believe the next step is for me to recommend to publish this article :)

There are two minor issues I came across to address:

1) The repository says CC-BY (https://github.com/villano-lab/galactic-spin-W1/blob/master/LICENSE) but this page says MIT https://galactic-spin-w1.readthedocs.io/en/latest/07_License.html. Does one or the other need to be updated?

2) There is a question mark in this reference that seems incorrect:
![image](https://github.com/openjournals/jose-reviews/assets/276007/e44f1e50-8997-49dc-9797-35d9c25bf8d9)

Other than that, this is a very high quality contribution and I'm happy to recommend to publish.

Address the two issues, make a new release, and report that version number here. I'll then update and start the process to publish.

Congrats!
