### 2023-05-17 #184 @Carreau
> @Carreau, would you mind providing an overall statement for your review also?

Sure. I was pleasantly surprised with the notebook provided by the author. It makes the topics easy to approach for non practitioners, without totally hiding the more complex details. I think this is great education material and I'm recommending for publishing.
