+++ docs/source/05_Citations.rst
@@ -8,7 +8,7 @@ How to cite
 
 If you decide to use this code, or if you want to add a reference to it, please cite the latest archived version,
 
-    Villano, A.N., Bergfalk, J., Hatami, R., Harris, K., Vititoe, F., and Johnston, J., 2022, The Data Behind Dark Matter: Exploring Galactic Rotation [Code, v2.0.0] [DOI:10.5281/zenodo.6588350]
+    Villano, A.N., Bergfalk, J., Hatami, R., Harris, K., Vititoe, F., and Johnston, J., 2023, The Data Behind Dark Matter: Exploring Galactic Rotation [Code, v3.0.2] [DOI:10.5281/zenodo.6588350]
 
 .. raw:: html
 
@@ -20,10 +20,10 @@ If you decide to use this code, or if you want to add a reference to it, please 
   @misc{villano_a_n_2022_6588350,
     author       = {Villano, A. N. and Harris, Kitty C. and Bergfalk, Judit and Hatami, Raphael and Vititoe, Francis and Johnston, Julia}, 
     title        = {The Data Behind Dark Matter: Exploring Galactic Rotation},
-    month        = {july},
-    year         = {2022},
+    month        = {may},
+    year         = {2023},
     publisher    = {Zenodo},
-    version      = {v3.0.1},
+    version      = {v3.0.2},
     doi          = {10.5281/zenodo.6588350},
     url          = {https://doi.org/10.5281/zenodo.6588350}
   }
