+++ docs/source/02_Modules.rst
@@ -91,9 +91,9 @@ Mass in a galaxy can be visualized by plotting its rotation curve. A galactic ro
 **Plotting measured velocities**
 
 .. image:: https://mybinder.org/badge_logo.svg
-   :target: https://mybinder.org/v2/gh/villano-lab/galactic-spin-W1/v3.0.2?labpath=binder%2F03_Measured_Data_Plotting.ipynb
+   :target: https://mybinder.org/v2/gh/villano-lab/galactic-spin-W1/v3.0.3?labpath=binder%2F03_Measured_Data_Plotting.ipynb
 
-`See module 03 on GitHub <https://github.com/villano-lab/galactic-spin-W1/blob/v3.0.2/binder/03_Measured_Data_Plotting.ipynb>`_
+`See module 03 on GitHub <https://github.com/villano-lab/galactic-spin-W1/blob/v3.0.3/binder/03_Measured_Data_Plotting.ipynb>`_
 
 First step in understanding a rotation curve is to plot it. Comparing the rotation curves of multiple galaxies is a good exercise to visualize the motions of stars and gas. The shapes of curves reveal the mass distribution in a galaxy. For example, higher velocity measurements in the central region indicates a supermassive black hole at the center of that galaxy. A flat rotation curve suggests the presence of a Dark Matter halo. 
 
@@ -106,9 +106,9 @@ First step in understanding a rotation curve is to plot it. Comparing the rotati
 **Plotting the components of galactic rotation curves (4 galaxies)**
 
 .. image:: https://mybinder.org/badge_logo.svg
-   :target: https://mybinder.org/v2/gh/villano-lab/galactic-spin-W1/v3.0.2?labpath=binder%2F04_Plotting_Rotation_Curves.ipynb 
+   :target: https://mybinder.org/v2/gh/villano-lab/galactic-spin-W1/v3.0.3?labpath=binder%2F04_Plotting_Rotation_Curves.ipynb 
 
-`See module 04 on GitHub <https://github.com/villano-lab/galactic-spin-W1/blob/v3.0.2/binder/04_Plotting_Rotation_Curves.ipynb>`_
+`See module 04 on GitHub <https://github.com/villano-lab/galactic-spin-W1/blob/v3.0.3/binder/04_Plotting_Rotation_Curves.ipynb>`_
 
 Theoretical rotation curves are plotted using the velocities of each component of the galaxy. The velocities of the bulge, disk and gas are calculated from luminosities, surface brightness profiles, surface density profiles or mass models but these calculations are beyond the scope of this workshop. For this reason, the velocities of the three components are imported into the notebook. On the contrary, the rotation curve of the central black hole (point-mass rotation curve - yet another type of rotation curve) and the dark matter halo can be easily produced. Adding all components, the total velocity of only light matter can then be compared to the total velocity that includes the dark matter component. Is it possible to fit both curves to the measured data?
 
@@ -121,9 +121,9 @@ Theoretical rotation curves are plotted using the velocities of each component o
 **Interactive widget to visualize the components of NGC 5533**
 
 .. image:: https://mybinder.org/badge_logo.svg
-   :target: https://mybinder.org/v2/gh/villano-lab/galactic-spin-W1/v3.0.2?labpath=binder%2F05_Widget_NGC5533_All_Components.ipynb 
+   :target: https://mybinder.org/v2/gh/villano-lab/galactic-spin-W1/v3.0.3?labpath=binder%2F05_Widget_NGC5533_All_Components.ipynb 
 
[…]
