### 2023-05-18 #184 @villaa
@moorepants we have fixed those issues and we are on v3.0.3 now. Thanks!
