### 2023-05-18 #184 @moorepants
@villaa can you make a zenodo archive of 3.0.3 as described above and then report the doi here.
