### 2023-05-18 #184 @nuclearGoblin
That DOI looks correct for me from campus internet but I get the login page anywhere else. We discussed that one before and had decided to cite it as a traditional book without a DOI because of the login page, but I don't think I was able to get to the site at the time. Here's what it looks like on my end:

![image](https://github.com/openjournals/jose-reviews/assets/45541156/1f064086-03e5-4092-bb6b-5b22c988bc48)

With that in mind should we add the DOI in?
