### 2023-05-18 #184 @moorepants
Very odd that a DOI would lead to a non-public page. If it is a book I'd just leave the DOI off, as books don't traditionally have DOIs.
