### 2023-05-19 #184 @moorepants
The DOI isn't that big of a deal, so if you want to archive to zenodo without fixing that, then do so.
