### 2023-05-19 #184 @villaa
> The DOI isn't that big of a deal, so if you want to archive to zenodo without fixing that, then do so.

Ok, @moorepants we are on version v3.0.3 and it has been added to our zenodo tag too. We are ready to go I think.
