### 2023-05-30 #184 @moorepants
Hi @labarba, just a friendly ping. This paper should be ready for publishing.
