+++ paper/references.bib
@@ -338,7 +338,7 @@ archivePrefix = {arXiv},
   year         = {2022},
   publisher    = {Zenodo},
   version      = {v1.0.2},
-  doi          = {DOI:10.5281/zenodo.6588350},
+  doi          = {10.5281/zenodo.6588350},
   url          = {https://zenodo.org/record/6588350},
   howpublished={The code can be found under \url{https://github.com/villano-lab/galactic-spin-W1/}.}
 }
