### 2023-08-07 #184 @labarba
Hi 👋
This should fix the DOI pattern error: https://github.com/villano-lab/galactic-spin-W1/pull/38
and this is a capitalization fix: https://github.com/villano-lab/galactic-spin-W1/pull/39
