### 2023-08-08 038e644
Create CITATION.cff
Copied from editorialbot [here](https://github.com/openjournals/jose-reviews/issues/184#issuecomment-1669516582)
