+++ CITATION.cff
@@ -1,0 +1,50 @@
+cff-version: "1.2.0"
+authors:
+- family-names: Villano
+  given-names: A. N.
+  orcid: "https://orcid.org/0000-0002-3893-7259"
+- family-names: Harris
+  given-names: Kitty C.
+  orcid: "https://orcid.org/0000-0001-5406-8367"
+- family-names: Bergfalk
+  given-names: Judit
+  orcid: "https://orcid.org/0000-0003-1662-0768"
+- family-names: Hatami
+  given-names: Raphael
+- family-names: Vititoe
+  given-names: Francis
+- family-names: Johnston
+  given-names: Julia
+doi: 10.5281/zenodo.7951874
+message: If you use this software, please cite our article in the
+  Journal of Open Source Software.
+preferred-citation:
+  authors:
+  - family-names: Villano
+    given-names: A. N.
+    orcid: "https://orcid.org/0000-0002-3893-7259"
+  - family-names: Harris
+    given-names: Kitty C.
+    orcid: "https://orcid.org/0000-0001-5406-8367"
+  - family-names: Bergfalk
+    given-names: Judit
+    orcid: "https://orcid.org/0000-0003-1662-0768"
+  - family-names: Hatami
+    given-names: Raphael
+  - family-names: Vititoe
+    given-names: Francis
+  - family-names: Johnston
+    given-names: Julia
+  date-published: 2023-08-08
+  doi: 10.21105/jose.00184
+  issn: 2577-3569
+  issue: 66
+  journal: Journal of Open Source Education
+  publisher:
+    name: Open Journals
+  start: 184
+  title: "The Data Behind Dark Matter: Exploring Galactic Rotation"
+  type: article
+  url: "https://jose.theoj.org/papers/10.21105/jose.00184"
+  volume: 6
+title: "The Data Behind Dark Matter: Exploring Galactic Rotation"
